@@ -4,7 +4,7 @@ version: 0.1.11
 homepage: https://famedly.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   http: ^0.13.0
@@ -22,6 +22,7 @@ dependencies:
   ffi: ^1.0.0
   js: ^0.6.3
   slugify: ^2.0.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   pedantic: ^1.11.0
