@@ -29,32 +29,32 @@ import 'utils/olm_session.dart';
 
 class OlmManager {
   final Encryption encryption;
-  Client get client => encryption.client;
-  olm.Account _olmAccount;
+  Client? get client => encryption.client;
+  olm.Account? _olmAccount;
 
   /// Returns the base64 encoded keys to store them in a store.
   /// This String should **never** leave the device!
-  String get pickledOlmAccount =>
-      enabled ? _olmAccount.pickle(client.userID) : null;
-  String get fingerprintKey =>
-      enabled ? json.decode(_olmAccount.identity_keys())['ed25519'] : null;
-  String get identityKey =>
-      enabled ? json.decode(_olmAccount.identity_keys())['curve25519'] : null;
+  String? get pickledOlmAccount =>
+      enabled ? _olmAccount!.pickle(client!.userID!) : null;
+  String? get fingerprintKey =>
+      enabled ? json.decode(_olmAccount!.identity_keys())['ed25519'] : null;
+  String? get identityKey =>
+      enabled ? json.decode(_olmAccount!.identity_keys())['curve25519'] : null;
 
   bool get enabled => _olmAccount != null;
 
   OlmManager(this.encryption);
 
   /// A map from Curve25519 identity keys to existing olm sessions.
-  Map<String, List<OlmSession>> get olmSessions => _olmSessions;
-  final Map<String, List<OlmSession>> _olmSessions = {};
+  Map<String?, List<OlmSession>> get olmSessions => _olmSessions;
+  final Map<String?, List<OlmSession>> _olmSessions = {};
 
-  Future<void> init(String olmAccount) async {
+  Future<void> init(String? olmAccount) async {
     if (olmAccount == null) {
       try {
         await olm.init();
         _olmAccount = olm.Account();
-        _olmAccount.create();
+        _olmAccount!.create();
         if (await uploadKeys(uploadDeviceKeys: true, updateDatabase: false) ==
             false) {
           throw ('Upload key failed');
@@ -68,7 +68,7 @@ class OlmManager {
       try {
         await olm.init();
         _olmAccount = olm.Account();
-        _olmAccount.unpickle(client.userID, olmAccount);
+        _olmAccount!.unpickle(client!.userID!, olmAccount);
       } catch (_) {
         _olmAccount?.free();
         _olmAccount = null;
@@ -81,21 +81,21 @@ class OlmManager {
   /// json.
   Map<String, dynamic> signJson(Map<String, dynamic> payload) {
     if (!enabled) throw ('Encryption is disabled');
-    final Map<String, dynamic> unsigned = payload['unsigned'];
-    final Map<String, dynamic> signatures = payload['signatures'];
+    final Map<String, dynamic>? unsigned = payload['unsigned'];
+    final Map<String, dynamic>? signatures = payload['signatures'];
     payload.remove('unsigned');
     payload.remove('signatures');
     final canonical = canonicalJson.encode(payload);
-    final signature = _olmAccount.sign(String.fromCharCodes(canonical));
+    final signature = _olmAccount!.sign(String.fromCharCodes(canonical));
     if (signatures != null) {
       payload['signatures'] = signatures;
     } else {
       payload['signatures'] = <String, dynamic>{};
     }
-    if (!payload['signatures'].containsKey(client.userID)) {
-      payload['signatures'][client.userID] = <String, dynamic>{};
+    if (!payload['signatures'].containsKey(client!.userID)) {
+      payload['signatures'][client!.userID] = <String, dynamic>{};
     }
-    payload['signatures'][client.userID]['ed25519:${client.deviceID}'] =
+    payload['signatures'][client!.userID]['ed25519:${client!.deviceID}'] =
         signature;
     if (unsigned != null) {
       payload['unsigned'] = unsigned;
@@ -104,7 +104,7 @@ class OlmManager {
   }
 
   String signString(String s) {
-    return _olmAccount.sign(s);
+    return _olmAccount!.sign(s);
   }
 
   /// Checks the signature of a signed json object.
@@ -112,7 +112,7 @@ class OlmManager {
   bool checkJsonSignature(String key, Map<String, dynamic> signedJson,
       String userId, String deviceId) {
     if (!enabled) throw ('Encryption is disabled');
-    final Map<String, dynamic> signatures = signedJson['signatures'];
+    final Map<String, dynamic>? signatures = signedJson['signatures'];
     if (signatures == null || !signatures.containsKey(userId)) return false;
     signedJson.remove('unsigned');
     signedJson.remove('signatures');
@@ -139,9 +139,9 @@ class OlmManager {
   /// Generates new one time keys, signs everything and upload it to the server.
   Future<bool> uploadKeys({
     bool uploadDeviceKeys = false,
-    int oldKeyCount = 0,
+    int? oldKeyCount = 0,
     bool updateDatabase = true,
-    bool unusedFallbackKey = false,
+    bool? unusedFallbackKey = false,
   }) async {
     if (!enabled) {
       return true;
@@ -154,27 +154,27 @@ class OlmManager {
 
     try {
       final signedOneTimeKeys = <String, dynamic>{};
-      int uploadedOneTimeKeysCount;
+      int? uploadedOneTimeKeysCount;
       if (oldKeyCount != null) {
         // check if we have OTKs that still need uploading. If we do, we don't try to generate new ones,
         // instead we try to upload the old ones first
         final oldOTKsNeedingUpload = json
-            .decode(_olmAccount.one_time_keys())['curve25519']
+            .decode(_olmAccount!.one_time_keys())['curve25519']
             .entries
             .length;
         // generate one-time keys
         // we generate 2/3rds of max, so that other keys people may still have can
         // still be used
         final oneTimeKeysCount =
-            (_olmAccount.max_number_of_one_time_keys() * 2 / 3).floor() -
+            (_olmAccount!.max_number_of_one_time_keys() * 2 / 3).floor() -
                 oldKeyCount -
                 oldOTKsNeedingUpload;
         if (oneTimeKeysCount > 0) {
-          _olmAccount.generate_one_time_keys(oneTimeKeysCount);
+          _olmAccount!.generate_one_time_keys(oneTimeKeysCount as int);
         }
-        uploadedOneTimeKeysCount = oneTimeKeysCount + oldOTKsNeedingUpload;
+        uploadedOneTimeKeysCount = oneTimeKeysCount + oldOTKsNeedingUpload as int?;
         final Map<String, dynamic> oneTimeKeys =
-            json.decode(_olmAccount.one_time_keys());
+            json.decode(_olmAccount!.one_time_keys());
 
         // now sign all the one-time keys
         for (final entry in oneTimeKeys['curve25519'].entries) {
@@ -189,8 +189,8 @@ class OlmManager {
       final signedFallbackKeys = <String, dynamic>{};
       if (encryption.isMinOlmVersion(3, 2, 0) && unusedFallbackKey == false) {
         // we don't have an unused fallback key uploaded....so let's change that!
-        _olmAccount.generate_fallback_key();
-        final fallbackKey = json.decode(_olmAccount.fallback_key());
+        _olmAccount!.generate_fallback_key();
+        final fallbackKey = json.decode(_olmAccount!.fallback_key());
         // now sign all the fallback keys
         for (final entry in fallbackKey['curve25519'].entries) {
           final key = entry.key;
@@ -206,8 +206,8 @@ class OlmManager {
       final keysContent = <String, dynamic>{
         if (uploadDeviceKeys)
           'device_keys': {
-            'user_id': client.userID,
-            'device_id': client.deviceID,
+            'user_id': client!.userID,
+            'device_id': client!.deviceID,
             'algorithms': [
               AlgorithmTypes.olmV1Curve25519AesSha2,
               AlgorithmTypes.megolmV1AesSha2
@@ -217,11 +217,11 @@ class OlmManager {
       };
       if (uploadDeviceKeys) {
         final Map<String, dynamic> keys =
-            json.decode(_olmAccount.identity_keys());
+            json.decode(_olmAccount!.identity_keys());
         for (final entry in keys.entries) {
           final algorithm = entry.key;
           final value = entry.value;
-          keysContent['device_keys']['keys']['$algorithm:${client.deviceID}'] =
+          keysContent['device_keys']['keys']['$algorithm:${client!.deviceID}'] =
               value;
         }
         keysContent['device_keys'] =
@@ -232,9 +232,9 @@ class OlmManager {
       // in case the app gets killed during upload or the upload fails due to bad network
       // we can still re-try later
       if (updateDatabase) {
-        await client.database?.updateClientKeys(pickledOlmAccount, client.id);
+        await client!.database?.updateClientKeys(pickledOlmAccount, client!.id);
       }
-      final response = await client.uploadKeys(
+      final response = await client!.uploadKeys(
         deviceKeys: uploadDeviceKeys
             ? MatrixDeviceKeys.fromJson(keysContent['device_keys'])
             : null,
@@ -242,9 +242,9 @@ class OlmManager {
         fallbackKeys: signedFallbackKeys,
       );
       // mark the OTKs as published and save that to datbase
-      _olmAccount.mark_keys_as_published();
+      _olmAccount!.mark_keys_as_published();
       if (updateDatabase) {
-        await client.database?.updateClientKeys(pickledOlmAccount, client.id);
+        await client!.database?.updateClientKeys(pickledOlmAccount, client!.id);
       }
       return (uploadedOneTimeKeysCount != null &&
               response['signed_curve25519'] == uploadedOneTimeKeysCount) ||
@@ -255,7 +255,7 @@ class OlmManager {
   }
 
   void handleDeviceOneTimeKeysCount(
-      Map<String, int> countJson, List<String> unusedFallbackKeyTypes) {
+      Map<String, int>? countJson, List<String>? unusedFallbackKeyTypes) {
     if (!enabled) {
       return;
     }
@@ -275,10 +275,10 @@ class OlmManager {
     }
 
     // Only upload keys if they are less than half of the max or we have no unused fallback key
-    if (keyCount < (_olmAccount.max_number_of_one_time_keys() / 2) ||
+    if (keyCount < (_olmAccount!.max_number_of_one_time_keys() / 2) ||
         !unusedFallbackKey) {
       uploadKeys(
-        oldKeyCount: keyCount < (_olmAccount.max_number_of_one_time_keys() / 2)
+        oldKeyCount: keyCount < (_olmAccount!.max_number_of_one_time_keys() / 2)
             ? keyCount
             : null,
         unusedFallbackKey: haveFallbackKeys ? unusedFallbackKey : null,
@@ -288,24 +288,24 @@ class OlmManager {
 
   Future<void> storeOlmSession(OlmSession session) async {
     _olmSessions[session.identityKey] ??= <OlmSession>[];
-    final ix = _olmSessions[session.identityKey]
+    final ix = _olmSessions[session.identityKey]!
         .indexWhere((s) => s.sessionId == session.sessionId);
     if (ix == -1) {
       // add a new session
-      _olmSessions[session.identityKey].add(session);
+      _olmSessions[session.identityKey]!.add(session);
     } else {
       // update an existing session
-      _olmSessions[session.identityKey][ix] = session;
+      _olmSessions[session.identityKey]![ix] = session;
     }
-    if (client.database == null) {
+    if (client!.database == null) {
       return;
     }
-    await client.database.storeOlmSession(
-        client.id,
+    await client!.database!.storeOlmSession(
+        client!.id,
         session.identityKey,
         session.sessionId,
         session.pickledSession,
-        session.lastReceived.millisecondsSinceEpoch);
+        session.lastReceived!.millisecondsSinceEpoch);
   }
 
   ToDeviceEvent _decryptToDeviceEvent(ToDeviceEvent event) {
@@ -316,38 +316,38 @@ class OlmManager {
     if (content.algorithm != AlgorithmTypes.olmV1Curve25519AesSha2) {
       throw DecryptException(DecryptException.unknownAlgorithm);
     }
-    if (!content.ciphertextOlm.containsKey(identityKey)) {
+    if (!content.ciphertextOlm!.containsKey(identityKey)) {
       throw DecryptException(DecryptException.isntSentForThisDevice);
     }
-    String plaintext;
+    String? plaintext;
     final senderKey = content.senderKey;
-    final body = content.ciphertextOlm[identityKey].body;
-    final type = content.ciphertextOlm[identityKey].type;
+    final body = content.ciphertextOlm![identityKey!]!.body;
+    final type = content.ciphertextOlm![identityKey!]!.type;
     if (type != 0 && type != 1) {
       throw DecryptException(DecryptException.unknownMessageType);
     }
-    final device = client.userDeviceKeys[event.sender]?.deviceKeys?.values
-        ?.firstWhere((d) => d.curve25519Key == senderKey, orElse: () => null);
+    final device = client!.userDeviceKeys[event.sender]?.deviceKeys?.values
+        ?.firstWhere((d) => d!.curve25519Key == senderKey, orElse: () => null);
     final existingSessions = olmSessions[senderKey];
-    final updateSessionUsage = ([OlmSession session]) => runInRoot(() async {
+    final Future<Null> Function([OlmSession]) updateSessionUsage = ([OlmSession? session]) => runInRoot(() async {
           if (session != null) {
             session.lastReceived = DateTime.now();
             await storeOlmSession(session);
           }
           if (device != null) {
             device.lastActive = DateTime.now();
-            await client.database?.setLastActiveUserDeviceKey(
-                device.lastActive.millisecondsSinceEpoch,
-                client.id,
+            await client!.database?.setLastActiveUserDeviceKey(
+                device.lastActive!.millisecondsSinceEpoch,
+                client!.id,
                 device.userId,
                 device.deviceId);
           }
         });
     if (existingSessions != null) {
       for (final session in existingSessions) {
-        if (type == 0 && session.session.matches_inbound(body) == true) {
+        if (type == 0 && session.session!.matches_inbound(body) == true) {
           try {
-            plaintext = session.session.decrypt(type, body);
+            plaintext = session.session!.decrypt(type, body);
           } catch (e) {
             // The message was encrypted during this session, but is unable to decrypt
             throw DecryptException(
@@ -357,7 +357,7 @@ class OlmManager {
           break;
         } else if (type == 1) {
           try {
-            plaintext = session.session.decrypt(type, body);
+            plaintext = session.session!.decrypt(type, body);
             updateSessionUsage(session);
             break;
           } catch (_) {
@@ -373,12 +373,12 @@ class OlmManager {
     if (plaintext == null) {
       final newSession = olm.Session();
       try {
-        newSession.create_inbound_from(_olmAccount, senderKey, body);
-        _olmAccount.remove_one_time_keys(newSession);
-        client.database?.updateClientKeys(pickledOlmAccount, client.id);
+        newSession.create_inbound_from(_olmAccount!, senderKey, body);
+        _olmAccount!.remove_one_time_keys(newSession);
+        client!.database?.updateClientKeys(pickledOlmAccount, client!.id);
         plaintext = newSession.decrypt(type, body);
         runInRoot(() => storeOlmSession(OlmSession(
-              key: client.userID,
+              key: client!.userID,
               identityKey: senderKey,
               sessionId: newSession.session_id(),
               session: newSession,
@@ -396,7 +396,7 @@ class OlmManager {
       throw DecryptException(DecryptException.senderDoesntMatch);
     }
     if (plainContent.containsKey('recipient') &&
-        plainContent['recipient'] != client.userID) {
+        plainContent['recipient'] != client!.userID) {
       throw DecryptException(DecryptException.recipientDoesntMatch);
     }
     if (plainContent['recipient_keys'] is Map &&
@@ -412,30 +412,30 @@ class OlmManager {
     );
   }
 
-  Future<List<OlmSession>> getOlmSessionsFromDatabase(String senderKey) async {
-    if (client.database == null) {
+  Future<List<OlmSession>> getOlmSessionsFromDatabase(String? senderKey) async {
+    if (client!.database == null) {
       return [];
     }
-    final olmSessions = await client.database
-        .getOlmSessions(client.id, senderKey, client.userID);
+    final olmSessions = await client!.database!
+        .getOlmSessions(client!.id, senderKey, client!.userID);
     return olmSessions.where((sess) => sess.isValid).toList();
   }
 
   Future<void> getOlmSessionsForDevicesFromDatabase(
-      List<String> senderKeys) async {
-    if (client.database == null) {
+      List<String?> senderKeys) async {
+    if (client!.database == null) {
       return;
     }
-    final rows = await client.database.getOlmSessionsForDevices(
-      client.id,
+    final rows = await client!.database!.getOlmSessionsForDevices(
+      client!.id,
       senderKeys,
-      client.userID,
+      client!.userID,
     );
-    final res = <String, List<OlmSession>>{};
+    final res = <String?, List<OlmSession>>{};
     for (final sess in rows) {
       res[sess.identityKey] ??= <OlmSession>[];
       if (sess.isValid) {
-        res[sess.identityKey].add(sess);
+        res[sess.identityKey]!.add(sess);
       }
     }
     for (final entry in res.entries) {
@@ -443,8 +443,8 @@ class OlmManager {
     }
   }
 
-  Future<List<OlmSession>> getOlmSessions(String senderKey,
-      {bool getFromDb = true}) async {
+  Future<List<OlmSession>> getOlmSessions(String? senderKey,
+      {bool? getFromDb = true}) async {
     var sess = olmSessions[senderKey];
     if ((getFromDb ?? true) && (sess == null || sess.isEmpty)) {
       final sessions = await getOlmSessionsFromDatabase(senderKey);
@@ -457,19 +457,19 @@ class OlmManager {
       return [];
     }
     sess.sort((a, b) => a.lastReceived == b.lastReceived
-        ? a.sessionId.compareTo(b.sessionId)
-        : b.lastReceived.compareTo(a.lastReceived));
+        ? a.sessionId!.compareTo(b.sessionId!)
+        : b.lastReceived!.compareTo(a.lastReceived!));
     return sess;
   }
 
   final Map<String, DateTime> _restoredOlmSessionsTime = {};
 
-  Future<void> restoreOlmSession(String userId, String senderKey) async {
-    if (!client.userDeviceKeys.containsKey(userId)) {
+  Future<void> restoreOlmSession(String? userId, String senderKey) async {
+    if (!client!.userDeviceKeys.containsKey(userId)) {
       return;
     }
-    final device = client.userDeviceKeys[userId].deviceKeys.values
-        .firstWhere((d) => d.curve25519Key == senderKey, orElse: () => null);
+    final device = client!.userDeviceKeys[userId]!.deviceKeys.values
+        .firstWhere((d) => d!.curve25519Key == senderKey, orElse: () => null);
     if (device == null) {
       return;
     }
@@ -478,12 +478,12 @@ class OlmManager {
     if (_restoredOlmSessionsTime.containsKey(mapKey) &&
         DateTime.now()
             .subtract(Duration(hours: 1))
-            .isBefore(_restoredOlmSessionsTime[mapKey])) {
+            .isBefore(_restoredOlmSessionsTime[mapKey]!)) {
       return;
     }
     _restoredOlmSessionsTime[mapKey] = DateTime.now();
     await startOutgoingOlmSessions([device]);
-    await client.sendToDeviceEncrypted([device], EventTypes.Dummy, {});
+    await client!.sendToDeviceEncrypted([device], EventTypes.Dummy, {});
   }
 
   Future<ToDeviceEvent> decryptToDeviceEvent(ToDeviceEvent event) async {
@@ -507,7 +507,7 @@ class OlmManager {
       return _decryptToDeviceEvent(event);
     } catch (_) {
       // okay, the thing errored while decrypting. It is safe to assume that the olm session is corrupt and we should generate a new one
-      if (client.enableE2eeRecovery) {
+      if (client!.enableE2eeRecovery) {
         // ignore: unawaited_futures
         runInRoot(() => restoreOlmSession(event.senderId, senderKey));
       }
@@ -515,27 +515,27 @@ class OlmManager {
     }
   }
 
-  Future<void> startOutgoingOlmSessions(List<DeviceKeys> deviceKeys) async {
+  Future<void> startOutgoingOlmSessions(List<DeviceKeys?> deviceKeys) async {
     Logs().v(
         '[OlmManager] Starting session with ${deviceKeys.length} devices...');
-    final requestingKeysFrom = <String, Map<String, String>>{};
+    final requestingKeysFrom = <String, Map<String?, String>>{};
     for (final device in deviceKeys) {
-      if (requestingKeysFrom[device.userId] == null) {
+      if (requestingKeysFrom[device!.userId] == null) {
         requestingKeysFrom[device.userId] = {};
       }
-      requestingKeysFrom[device.userId][device.deviceId] = 'signed_curve25519';
+      requestingKeysFrom[device.userId]![device.deviceId] = 'signed_curve25519';
     }
 
-    final response = await client.claimKeys(requestingKeysFrom, timeout: 10000);
+    final response = await client!.claimKeys(requestingKeysFrom as Map<String, Map<String, String>>, timeout: 10000);
 
     for (final userKeysEntry in response.oneTimeKeys.entries) {
       final userId = userKeysEntry.key;
       for (final deviceKeysEntry in userKeysEntry.value.entries) {
         final deviceId = deviceKeysEntry.key;
         final fingerprintKey =
-            client.userDeviceKeys[userId].deviceKeys[deviceId].ed25519Key;
+            client!.userDeviceKeys[userId]!.deviceKeys[deviceId]!.ed25519Key;
         final identityKey =
-            client.userDeviceKeys[userId].deviceKeys[deviceId].curve25519Key;
+            client!.userDeviceKeys[userId]!.deviceKeys[deviceId]!.curve25519Key;
         for (final Map<String, dynamic> deviceKey
             in deviceKeysEntry.value.values) {
           if (!deviceKey.checkJsonSignature(fingerprintKey, userId, deviceId)) {
@@ -544,9 +544,9 @@ class OlmManager {
           Logs().v('[OlmManager] Starting session with $userId:$deviceId');
           final session = olm.Session();
           try {
-            session.create_outbound(_olmAccount, identityKey, deviceKey['key']);
+            session.create_outbound(_olmAccount!, identityKey!, deviceKey['key']);
             await storeOlmSession(OlmSession(
-              key: client.userID,
+              key: client!.userID,
               identityKey: identityKey,
               sessionId: session.session_id(),
               session: session,
@@ -564,8 +564,8 @@ class OlmManager {
   }
 
   Future<Map<String, dynamic>> encryptToDeviceMessagePayload(
-      DeviceKeys device, String type, Map<String, dynamic> payload,
-      {bool getFromDb}) async {
+      DeviceKeys device, String? type, Map<String, dynamic>? payload,
+      {bool? getFromDb}) async {
     final sess =
         await getOlmSessions(device.curve25519Key, getFromDb: getFromDb);
     if (sess.isEmpty) {
@@ -574,21 +574,21 @@ class OlmManager {
     final fullPayload = {
       'type': type,
       'content': payload,
-      'sender': client.userID,
+      'sender': client!.userID,
       'keys': {'ed25519': fingerprintKey},
       'recipient': device.userId,
       'recipient_keys': {'ed25519': device.ed25519Key},
     };
-    final encryptResult = sess.first.session.encrypt(json.encode(fullPayload));
+    final encryptResult = sess.first.session!.encrypt(json.encode(fullPayload));
     await storeOlmSession(sess.first);
-    if (client.database != null) {
+    if (client!.database != null) {
       // ignore: unawaited_futures
-      runInRoot(() => client.database.setLastSentMessageUserDeviceKey(
+      runInRoot(() => client!.database!.setLastSentMessageUserDeviceKey(
           json.encode({
             'type': type,
             'content': payload,
           }),
-          client.id,
+          client!.id,
           device.userId,
           device.deviceId));
     }
@@ -605,28 +605,28 @@ class OlmManager {
   }
 
   Future<Map<String, dynamic>> encryptToDeviceMessage(
-      List<DeviceKeys> deviceKeys,
-      String type,
-      Map<String, dynamic> payload) async {
-    final data = <String, Map<String, Map<String, dynamic>>>{};
+      List<DeviceKeys?> deviceKeys,
+      String? type,
+      Map<String, dynamic>? payload) async {
+    final data = <String, Map<String?, Map<String, dynamic>>>{};
     // first check if any of our sessions we want to encrypt for are in the database
-    if (client.database != null) {
+    if (client!.database != null) {
       await getOlmSessionsForDevicesFromDatabase(
-          deviceKeys.map((d) => d.curve25519Key).toList());
+          deviceKeys.map((d) => d!.curve25519Key).toList());
     }
     final deviceKeysWithoutSession = List<DeviceKeys>.from(deviceKeys);
     deviceKeysWithoutSession.removeWhere((DeviceKeys deviceKeys) =>
         olmSessions.containsKey(deviceKeys.curve25519Key) &&
-        olmSessions[deviceKeys.curve25519Key].isNotEmpty);
+        olmSessions[deviceKeys.curve25519Key]!.isNotEmpty);
     if (deviceKeysWithoutSession.isNotEmpty) {
       await startOutgoingOlmSessions(deviceKeysWithoutSession);
     }
     for (final device in deviceKeys) {
-      if (!data.containsKey(device.userId)) {
+      if (!data.containsKey(device!.userId)) {
         data[device.userId] = {};
       }
       try {
-        data[device.userId][device.deviceId] =
+        data[device.userId]![device.deviceId] =
             await encryptToDeviceMessagePayload(device, type, payload,
                 getFromDb: false);
       } catch (e, s) {
@@ -641,26 +641,26 @@ class OlmManager {
     if (event.type == EventTypes.Dummy) {
       // We receive dan encrypted m.dummy. This means that the other end was not able to
       // decrypt our last message. So, we re-send it.
-      if (event.encryptedContent == null || client.database == null) {
+      if (event.encryptedContent == null || client!.database == null) {
         return;
       }
-      final device = client.getUserDeviceKeysByCurve25519Key(
-          event.encryptedContent.tryGet<String>('sender_key') ?? '');
+      final device = client!.getUserDeviceKeysByCurve25519Key(
+          event.encryptedContent!.tryGet<String>('sender_key') ?? '');
       if (device == null) {
         return; // device not found
       }
       Logs().v(
           '[OlmManager] Device ${device.userId}:${device.deviceId} generated a new olm session, replaying last sent message...');
-      final lastSentMessageRes = await client.database
+      final lastSentMessageRes = await client!.database!
           .getLastSentMessageUserDeviceKey(
-              client.id, device.userId, device.deviceId);
+              client!.id, device.userId, device.deviceId);
       if (lastSentMessageRes.isEmpty ||
           (lastSentMessageRes.first?.isEmpty ?? true)) {
         return;
       }
-      final lastSentMessage = json.decode(lastSentMessageRes.first);
+      final lastSentMessage = json.decode(lastSentMessageRes.first!);
       // okay, time to send the message!
-      await client.sendToDeviceEncrypted(
+      await client!.sendToDeviceEncrypted(
           [device], lastSentMessage['type'], lastSentMessage['content']);
     }
   }
