@@ -41,7 +41,7 @@ import 'utils/to_device_event.dart';
 import 'utils/uia_request.dart';
 import 'utils/multilock.dart';
 
-typedef RoomSorter = int Function(Room a, Room b);
+typedef RoomSorter = int Function(Room? a, Room? b);
 
 enum LoginState { logged, loggedOut }
 
@@ -55,36 +55,36 @@ extension TrailingSlash on Uri {
 /// [Matrix](https://matrix.org) homeserver and is the entry point for this
 /// SDK.
 class Client extends MatrixApi {
-  int _id;
+  late int _id;
 
   // Keeps track of the currently ongoing syncRequest
   // in case we want to cancel it.
-  int _currentSyncId;
+  int? _currentSyncId;
 
   int get id => _id;
 
-  final FutureOr<DatabaseApi> Function(Client) databaseBuilder;
-  final FutureOr<DatabaseApi> Function(Client) legacyDatabaseBuilder;
-  final FutureOr<void> Function(Client) databaseDestroyer;
-  final FutureOr<void> Function(Client) legacyDatabaseDestroyer;
-  DatabaseApi _database;
+  final FutureOr<DatabaseApi>? Function(Client)? databaseBuilder;
+  final FutureOr<DatabaseApi> Function(Client)? legacyDatabaseBuilder;
+  final FutureOr<void> Function(Client)? databaseDestroyer;
+  final FutureOr<void> Function(Client)? legacyDatabaseDestroyer;
+  DatabaseApi? _database;
 
-  DatabaseApi get database => _database;
+  DatabaseApi? get database => _database;
 
   bool enableE2eeRecovery;
 
   @deprecated
   MatrixApi get api => this;
 
-  Encryption encryption;
+  Encryption? encryption;
 
-  Set<KeyVerificationMethod> verificationMethods;
+  Set<KeyVerificationMethod>? verificationMethods;
 
-  Set<String> importantStateEvents;
+  Set<String>? importantStateEvents;
 
-  Set<String> roomPreviewLastEvents;
+  Set<String>? roomPreviewLastEvents;
 
-  Set<String> supportedLoginTypes;
+  Set<String>? supportedLoginTypes;
 
   int sendMessageTimeoutSeconds;
 
@@ -95,18 +95,18 @@ class Client extends MatrixApi {
   bool mxidLocalPartFallback = true;
 
   // For CommandsClientExtension
-  final Map<String, FutureOr<String> Function(CommandArgs)> commands = {};
+  final Map<String, FutureOr<String?> Function(CommandArgs)> commands = {};
   final Filter syncFilter;
 
-  String syncFilterId;
+  String? syncFilterId;
 
   final Future<R> Function<Q, R>(FutureOr<R> Function(Q), Q,
-      {String debugLabel}) compute;
+      {String? debugLabel})? compute;
 
   Future<T> runInBackground<T, U>(
       FutureOr<T> Function(U arg) function, U arg) async {
     if (compute != null) {
-      return await compute(function, arg);
+      return await compute!(function, arg);
     }
     return await function(arg);
   }
@@ -155,7 +155,7 @@ class Client extends MatrixApi {
     this.legacyDatabaseDestroyer,
     this.enableE2eeRecovery = false,
     this.verificationMethods,
-    http.Client httpClient,
+    http.Client? httpClient,
     this.importantStateEvents,
     this.roomPreviewLastEvents,
     this.pinUnreadRooms = false,
@@ -163,8 +163,8 @@ class Client extends MatrixApi {
     this.requestHistoryOnLimitedTimeline = false,
     this.supportedLoginTypes,
     this.compute,
-    Filter syncFilter,
-    @deprecated bool debug,
+    Filter? syncFilter,
+    @deprecated bool? debug,
   })  : syncFilter = syncFilter ??
             Filter(
               room: RoomFilter(
@@ -177,7 +177,7 @@ class Client extends MatrixApi {
     supportedLoginTypes ??= {AuthenticationTypes.password};
     verificationMethods ??= <KeyVerificationMethod>{};
     importantStateEvents ??= {};
-    importantStateEvents.addAll([
+    importantStateEvents!.addAll([
       EventTypes.RoomName,
       EventTypes.RoomAvatar,
       EventTypes.Message,
@@ -187,7 +187,7 @@ class Client extends MatrixApi {
       EventTypes.RoomTombstone,
     ]);
     roomPreviewLastEvents ??= {};
-    roomPreviewLastEvents.addAll([
+    roomPreviewLastEvents!.addAll([
       EventTypes.Message,
       EventTypes.Encrypted,
       EventTypes.Sticker,
@@ -201,23 +201,23 @@ class Client extends MatrixApi {
   final String clientName;
 
   /// The Matrix ID of the current logged user.
-  String get userID => _userID;
-  String _userID;
+  String? get userID => _userID;
+  String? _userID;
 
   /// This points to the position in the synchronization history.
-  String prevBatch;
+  String? prevBatch;
 
   /// The device ID is an unique identifier for this device.
-  String get deviceID => _deviceID;
-  String _deviceID;
+  String? get deviceID => _deviceID;
+  String? _deviceID;
 
   /// The device name is a human readable identifier for this device.
-  String get deviceName => _deviceName;
-  String _deviceName;
+  String? get deviceName => _deviceName;
+  String? _deviceName;
 
   /// Returns the current login state.
-  LoginState get loginState => __loginState;
-  LoginState __loginState;
+  LoginState? get loginState => __loginState;
+  LoginState? __loginState;
   set _loginState(LoginState state) {
     __loginState = state;
     onLoginStateChanged.add(state);
@@ -226,11 +226,11 @@ class Client extends MatrixApi {
   bool isLogged() => accessToken != null;
 
   /// A list of all rooms the user is participating or invited.
-  List<Room> get rooms => _rooms;
-  List<Room> _rooms = [];
+  List<Room?> get rooms => _rooms;
+  List<Room?> _rooms = [];
 
   /// Whether this client supports end-to-end encryption using olm.
-  bool get encryptionEnabled => encryption != null && encryption.enabled;
+  bool get encryptionEnabled => encryption != null && encryption!.enabled;
 
   /// Whether this client is able to encrypt and decrypt files.
   bool get fileEncryptionEnabled => encryptionEnabled && true;
@@ -242,17 +242,17 @@ class Client extends MatrixApi {
   /// Wheather this session is unknown to others
   bool get isUnknownSession =>
       !userDeviceKeys.containsKey(userID) ||
-      !userDeviceKeys[userID].deviceKeys.containsKey(deviceID) ||
-      !userDeviceKeys[userID].deviceKeys[deviceID].signed;
+      !userDeviceKeys[userID]!.deviceKeys.containsKey(deviceID) ||
+      !userDeviceKeys[userID]!.deviceKeys[deviceID]!.signed;
 
   /// Warning! This endpoint is for testing only!
-  set rooms(List<Room> newList) {
+  set rooms(List<Room?> newList) {
     Logs().w('Warning! This endpoint is for testing only!');
     _rooms = newList;
   }
 
   /// Key/Value store of account data.
-  Map<String, BasicEvent> accountData = {};
+  Map<String?, BasicEvent> accountData = {};
 
   /// Presences of users by a given matrix ID
   Map<String, Presence> presences = {};
@@ -264,30 +264,30 @@ class Client extends MatrixApi {
     return '$clientName-$_transactionCounter-${DateTime.now().millisecondsSinceEpoch}';
   }
 
-  Room getRoomByAlias(String alias) {
+  Room? getRoomByAlias(String alias) {
     for (final room in rooms) {
-      if (room.canonicalAlias == alias) return room;
+      if (room!.canonicalAlias == alias) return room;
     }
     return null;
   }
 
-  Room getRoomById(String id) {
+  Room? getRoomById(String? id) {
     for (final room in rooms) {
-      if (room.id == id) return room;
+      if (room!.id == id) return room;
     }
     return null;
   }
 
-  Map<String, dynamic> get directChats =>
-      accountData['m.direct'] != null ? accountData['m.direct'].content : {};
+  Map<String?, dynamic> get directChats =>
+      accountData['m.direct'] != null ? accountData['m.direct']!.content : {};
 
   /// Returns the (first) room ID from the store which is a private chat with the user [userId].
   /// Returns null if there is none.
-  String getDirectChatFromUserId(String userId) {
+  String? getDirectChatFromUserId(String? userId) {
     if (accountData['m.direct'] != null &&
-        accountData['m.direct'].content[userId] is List<dynamic> &&
-        accountData['m.direct'].content[userId].length > 0) {
-      final potentialRooms = accountData['m.direct']
+        accountData['m.direct']!.content[userId!] is List<dynamic> &&
+        accountData['m.direct']!.content[userId].length > 0) {
+      final potentialRooms = accountData['m.direct']!
           .content[userId]
           .cast<String>()
           .map(getRoomById)
@@ -306,9 +306,9 @@ class Client extends MatrixApi {
       }
     }
     for (final room in rooms) {
-      if (room.membership == Membership.invite &&
+      if (room!.membership == Membership.invite &&
           room.getState(EventTypes.RoomMember, userID)?.senderId == userId &&
-          room.getState(EventTypes.RoomMember, userID).content['is_direct'] ==
+          room.getState(EventTypes.RoomMember, userID)!.content['is_direct'] ==
               true) {
         return room.id;
       }
@@ -322,7 +322,7 @@ class Client extends MatrixApi {
   ) async {
     try {
       final response = await http.get(
-          Uri.https(MatrixIdOrDomain.domain, '/.well-known/matrix/client'));
+          Uri.https(MatrixIdOrDomain.domain!, '/.well-known/matrix/client'));
       var respBody = response.body;
       try {
         respBody = utf8.decode(response.bodyBytes);
@@ -336,7 +336,7 @@ class Client extends MatrixApi {
       // provide a reasonable fallback.
       return DiscoveryInformation(
         mHomeserver: HomeserverInformation(
-            baseUrl: Uri.https(MatrixIdOrDomain.domain, '')),
+            baseUrl: Uri.https(MatrixIdOrDomain.domain!, '')),
       );
     }
   }
@@ -355,15 +355,15 @@ class Client extends MatrixApi {
   /// login types. Throws an exception if the server is not compatible with the
   /// client and sets [homeserver] to [homeserverUrl] if it is. Supports the
   /// types `Uri` and `String`.
-  Future<DiscoveryInformation> checkHomeserver(dynamic homeserverUrl,
+  Future<DiscoveryInformation?> checkHomeserver(dynamic homeserverUrl,
       {bool checkWellKnown = true}) async {
     try {
       homeserver =
           (homeserverUrl is Uri) ? homeserverUrl : Uri.parse(homeserverUrl);
-      homeserver = homeserver.stripTrailingSlash();
+      homeserver = homeserver!.stripTrailingSlash();
 
       // Look up well known
-      DiscoveryInformation wellKnown;
+      DiscoveryInformation? wellKnown;
       if (checkWellKnown) {
         try {
           wellKnown = await getWellknown();
@@ -381,8 +381,8 @@ class Client extends MatrixApi {
             versions.versions.toSet(), supportedVersions);
       }
 
-      final loginTypes = await getLoginFlows();
-      if (!loginTypes.any((f) => supportedLoginTypes.contains(f.type))) {
+      final loginTypes = await (getLoginFlows() as FutureOr<List<LoginFlow>>);
+      if (!loginTypes.any((f) => supportedLoginTypes!.contains(f.type))) {
         throw BadServerLoginTypesException(
             loginTypes.map((f) => f.type).toSet(), supportedLoginTypes);
       }
@@ -399,13 +399,13 @@ class Client extends MatrixApi {
   /// You have to call [checkHomeserver] first to set a homeserver.
   @override
   Future<RegisterResponse> register({
-    String username,
-    String password,
-    String deviceId,
-    String initialDeviceDisplayName,
-    bool inhibitLogin,
-    AuthenticationData auth,
-    AccountKind kind,
+    String? username,
+    String? password,
+    String? deviceId,
+    String? initialDeviceDisplayName,
+    bool? inhibitLogin,
+    AuthenticationData? auth,
+    AccountKind? kind,
   }) async {
     final response = await super.register(
       username: username,
@@ -441,17 +441,17 @@ class Client extends MatrixApi {
   @override
   Future<LoginResponse> login(
     LoginType type, {
-    AuthenticationIdentifier identifier,
-    String password,
-    String token,
-    String deviceId,
-    String initialDeviceDisplayName,
-    AuthenticationData auth,
-    @Deprecated('Deprecated in favour of identifier.') String user,
-    @Deprecated('Deprecated in favour of identifier.') String medium,
-    @Deprecated('Deprecated in favour of identifier.') String address,
+    AuthenticationIdentifier? identifier,
+    String? password,
+    String? token,
+    String? deviceId,
+    String? initialDeviceDisplayName,
+    AuthenticationData? auth,
+    @Deprecated('Deprecated in favour of identifier.') String? user,
+    @Deprecated('Deprecated in favour of identifier.') String? medium,
+    @Deprecated('Deprecated in favour of identifier.') String? address,
   }) async {
-    if (homeserver == null && user.isValidMatrixId) {
+    if (homeserver == null && user!.isValidMatrixId) {
       await checkHomeserver(user.domain);
     }
     final loginResp = await super.login(
@@ -517,14 +517,14 @@ class Client extends MatrixApi {
   Future<T> uiaRequestBackground<T>(
       Future<T> Function(AuthenticationData auth) request) {
     final completer = Completer<T>();
-    UiaRequest uia;
+    UiaRequest? uia;
     uia = UiaRequest(
       request: request,
       onUpdate: (state) {
         if (state == UiaRequestState.done) {
-          completer.complete(uia.result);
+          completer.complete(uia!.result);
         } else if (state == UiaRequestState.fail) {
-          completer.completeError(uia.error);
+          completer.completeError(uia!.error);
         } else {
           onUiaRequest.add(uia);
         }
@@ -535,14 +535,14 @@ class Client extends MatrixApi {
 
   /// Returns an existing direct room ID with this user or creates a new one.
   /// Returns null on error.
-  Future<String> startDirectChat(String mxid) async {
+  Future<String> startDirectChat(String? mxid) async {
     // Try to find an existing direct chat
     var roomId = getDirectChatFromUserId(mxid);
     if (roomId != null) return roomId;
 
     // Start a new direct chat
     roomId = await createRoom(
-      invite: [mxid],
+      invite: [mxid!],
       isDirect: true,
       preset: CreateRoomPreset.trustedPrivateChat,
     );
@@ -562,13 +562,13 @@ class Client extends MatrixApi {
   ///
   /// https://github.com/matrix-org/matrix-doc/blob/matthew/msc1772/proposals/1772-groups-as-rooms.md
   Future<String> createSpace({
-    String name,
-    String topic,
+    String? name,
+    String? topic,
     Visibility visibility = Visibility.public,
-    String spaceAliasName,
-    List<String> invite,
-    List<Invite3pid> invite3pid,
-    String roomVersion,
+    String? spaceAliasName,
+    List<String>? invite,
+    List<Invite3pid>? invite3pid,
+    String? roomVersion,
   }) =>
       createRoom(
         name: name,
@@ -590,7 +590,7 @@ class Client extends MatrixApi {
     if (rooms.isNotEmpty) {
       final profileSet = <Profile>{};
       for (final room in rooms) {
-        final user = room.getUserByMXIDSync(userID);
+        final user = room!.getUserByMXIDSync(userID);
         profileSet.add(Profile.fromJson(user.content));
       }
       if (profileSet.length == 1) return profileSet.first;
@@ -598,7 +598,7 @@ class Client extends MatrixApi {
     return getProfileFromUserId(userID);
   }
 
-  final Map<String, ProfileInformation> _profileCache = {};
+  final Map<String?, ProfileInformation> _profileCache = {};
 
   /// Get the combined profile information for this user.
   /// If [getFromRooms] is true then the profile will first be searched from the
@@ -607,12 +607,12 @@ class Client extends MatrixApi {
   /// If [cache] is true then
   /// the profile get cached for this session. Please note that then the profile may
   /// become outdated if the user changes the displayname or avatar in this session.
-  Future<Profile> getProfileFromUserId(String userId,
+  Future<Profile> getProfileFromUserId(String? userId,
       {bool cache = true, bool getFromRooms = true}) async {
     if (getFromRooms) {
       final room = rooms.firstWhere(
-          (Room room) =>
-              room
+          (Room? room) =>
+              room!
                   .getParticipants()
                   .indexWhere((User user) => user.id == userId) !=
               -1,
@@ -621,19 +621,19 @@ class Client extends MatrixApi {
         final user =
             room.getParticipants().firstWhere((User user) => user.id == userId);
         return Profile(
-            userId: userId,
+            userId: userId!,
             displayName: user.displayName,
             avatarUrl: user.avatarUrl);
       }
     }
     if (cache && _profileCache.containsKey(userId)) {
-      final profile = _profileCache[userId];
+      final profile = _profileCache[userId]!;
       return Profile(
-          userId: userId,
+          userId: userId!,
           displayName: profile.displayname,
           avatarUrl: profile.avatarUrl);
     }
-    final profile = await getUserProfile(userId);
+    final profile = await getUserProfile(userId!);
     _profileCache[userId] = profile;
     return Profile(
         userId: userId,
@@ -647,8 +647,8 @@ class Client extends MatrixApi {
       filter: '{"room":{"include_leave":true,"timeline":{"limit":10}}}',
       timeout: 0,
     );
-    if (syncResp.rooms.leave is Map<String, dynamic>) {
-      for (final entry in syncResp.rooms.leave.entries) {
+    if (syncResp.rooms!.leave is Map<String, dynamic>) {
+      for (final entry in syncResp.rooms!.leave!.entries) {
         final id = entry.key;
         final room = entry.value;
         final leftRoom = Room(
@@ -660,7 +660,7 @@ class Client extends MatrixApi {
                   <String, BasicRoomEvent>{},
         );
         if (room.timeline?.events != null) {
-          for (final event in room.timeline.events) {
+          for (final event in room.timeline!.events!) {
             leftRoom.setState(Event.fromMatrixEvent(
               event,
               leftRoom,
@@ -669,7 +669,7 @@ class Client extends MatrixApi {
           }
         }
         if (room.state != null) {
-          for (final event in room.state) {
+          for (final event in room.state!) {
             leftRoom.setState(Event.fromMatrixEvent(
               event,
               leftRoom,
@@ -687,12 +687,12 @@ class Client extends MatrixApi {
   /// and returns the mxc url as a string.
   @override
   Future<String> uploadContent(Uint8List file,
-      {String filename, String contentType}) async {
+      {String? filename, String? contentType}) async {
     final mxc = await super
         .uploadContent(file, filename: filename, contentType: contentType);
-    final storeable = database != null && file.length <= database.maxFileSize;
+    final storeable = database != null && file.length <= database!.maxFileSize;
     if (storeable) {
-      await database.storeFile(
+      await database!.storeFile(
           mxc, file, DateTime.now().millisecondsSinceEpoch);
     }
     return mxc;
@@ -704,31 +704,31 @@ class Client extends MatrixApi {
     String userId,
     String roomId,
     bool typing, {
-    int timeout,
+    int? timeout,
   }) async {
     await super.setTyping(userId, roomId, typing, timeout: timeout);
     final room = getRoomById(roomId);
     if (typing && room != null && encryptionEnabled && room.encrypted) {
       // ignore: unawaited_futures
-      encryption.keyManager.prepareOutboundGroupSession(roomId);
+      encryption!.keyManager!.prepareOutboundGroupSession(roomId);
     }
   }
 
   /// Uploads a new user avatar for this user.
   Future<void> setAvatar(MatrixFile file) async {
-    final uploadResp = await uploadContent(file.bytes, filename: file.name);
-    await setAvatarUrl(userID, Uri.parse(uploadResp));
+    final uploadResp = await uploadContent(file.bytes!, filename: file.name);
+    await setAvatarUrl(userID!, Uri.parse(uploadResp));
     return;
   }
 
   /// Returns the global push rules for the logged in user.
-  PushRuleSet get globalPushRules => accountData.containsKey('m.push_rules')
-      ? PushRuleSet.fromJson(accountData['m.push_rules'].content['global'])
+  PushRuleSet? get globalPushRules => accountData.containsKey('m.push_rules')
+      ? PushRuleSet.fromJson(accountData['m.push_rules']!.content['global'])
       : null;
 
   /// Returns the device push rules for the logged in user.
-  PushRuleSet get devicePushRules => accountData.containsKey('m.push_rules')
-      ? PushRuleSet.fromJson(accountData['m.push_rules'].content['device'])
+  PushRuleSet? get devicePushRules => accountData.containsKey('m.push_rules')
+      ? PushRuleSet.fromJson(accountData['m.push_rules']!.content['device'])
       : null;
 
   static const Set<String> supportedVersions = {'r0.5.0', 'r0.6.0'};
@@ -806,7 +806,7 @@ class Client extends MatrixApi {
 
   /// When the library calls an endpoint that needs UIA the `UiaRequest` is passed down this screen.
   /// The client can open a UIA prompt based on this.
-  final StreamController<UiaRequest> onUiaRequest =
+  final StreamController<UiaRequest?> onUiaRequest =
       StreamController.broadcast();
 
   /// How long should the app wait until it retrys the synchronisation after
@@ -815,12 +815,12 @@ class Client extends MatrixApi {
 
   @Deprecated('Use init() instead')
   void connect({
-    String newToken,
-    Uri newHomeserver,
-    String newUserID,
-    String newDeviceName,
-    String newDeviceID,
-    String newOlmAccount,
+    String? newToken,
+    Uri? newHomeserver,
+    String? newUserID,
+    String? newDeviceName,
+    String? newDeviceID,
+    String? newOlmAccount,
   }) =>
       init(
         newToken: newToken,
@@ -846,13 +846,13 @@ class Client extends MatrixApi {
   /// If one of [newToken], [newUserID], [newDeviceID], [newDeviceName] is set then
   /// all of them must be set! If you don't set them, this method will try to
   /// get them from the database.
-  Future<void> init({
-    String newToken,
-    Uri newHomeserver,
-    String newUserID,
-    String newDeviceName,
-    String newDeviceID,
-    String newOlmAccount,
+  Future<void>? init({
+    String? newToken,
+    Uri? newHomeserver,
+    String? newUserID,
+    String? newDeviceName,
+    String? newDeviceID,
+    String? newOlmAccount,
   }) async {
     if ((newToken != null ||
             newUserID != null ||
@@ -875,12 +875,12 @@ class Client extends MatrixApi {
       }
 
       if (databaseBuilder != null) {
-        _database ??= await databaseBuilder(this);
+        _database ??= await databaseBuilder!(this);
       }
 
-      String olmAccount;
+      String? olmAccount;
       if (database != null) {
-        final account = await database.getClient(clientName);
+        final account = await database!.getClient(clientName);
         if (account != null) {
           _id = account['client_id'];
           homeserver = Uri.parse(account['homeserver_url']);
@@ -938,7 +938,7 @@ class Client extends MatrixApi {
 
       if (database != null) {
         if (id != null) {
-          await database.updateClient(
+          await database!.updateClient(
             homeserver.toString(),
             accessToken,
             _userID,
@@ -949,7 +949,7 @@ class Client extends MatrixApi {
             id,
           );
         } else {
-          _id = await database.insertClient(
+          _id = (await (database!.insertClient(
             clientName,
             homeserver.toString(),
             accessToken,
@@ -958,18 +958,18 @@ class Client extends MatrixApi {
             _deviceName,
             prevBatch,
             encryption?.pickledOlmAccount,
-          );
+          ) as FutureOr<int?>))!;
         }
-        _userDeviceKeys = await database.getUserDeviceKeys(this);
-        _rooms = await database.getRoomList(this);
+        _userDeviceKeys = await database!.getUserDeviceKeys(this);
+        _rooms = await database!.getRoomList(this);
         _sortRooms();
-        accountData = await database.getAccountData(id);
+        accountData = await database!.getAccountData(id);
         presences.clear();
       }
       _initLock = false;
       _loginState = LoginState.logged;
       Logs().i(
-        'Successfully connected as ${userID.localpart} with ${homeserver.toString()}',
+        'Successfully connected as ${userID!.localpart} with ${homeserver.toString()}',
       );
       return _sync();
     } catch (e, s) {
@@ -999,8 +999,14 @@ class Client extends MatrixApi {
       _database = null;
     }
 
-    _id = accessToken = syncFilterId =
-        homeserver = _userID = _deviceID = _deviceName = prevBatch = null;
+    _id = 0;
+    accessToken = null;
+    syncFilterId = null;
+    homeserver = null;
+    _userID = null;
+    _deviceID = null;
+    _deviceName = null;
+    prevBatch = null; //null;
     _rooms = [];
     encryption?.dispose();
     encryption = null;
@@ -1010,14 +1016,14 @@ class Client extends MatrixApi {
       } catch (e, s) {
         Logs().e('Unable to close database', e, s);
       }
-      await databaseDestroyer(this);
+      await databaseDestroyer!(this);
       _database = null;
     }
     _loginState = LoginState.loggedOut;
   }
 
   bool _backgroundSync = true;
-  Future<void> _currentSync, _retryDelay = Future.value();
+  Future<void>? _currentSync, _retryDelay = Future.value();
 
   bool get syncPending => _currentSync != null;
 
@@ -1031,14 +1037,14 @@ class Client extends MatrixApi {
 
   /// Immediately start a sync and wait for completion.
   /// If there is an active sync already, wait for the active sync instead.
-  Future<void> oneShotSync() {
+  Future<void>? oneShotSync() {
     return _sync();
   }
 
-  Future<void> _sync() {
+  Future<void>? _sync() {
     if (_currentSync == null) {
       _currentSync = _innerSync();
-      _currentSync.whenComplete(() {
+      _currentSync!.whenComplete(() {
         _currentSync = null;
         if (_backgroundSync && isLogged() && !_disposed) {
           _sync();
@@ -1049,11 +1055,11 @@ class Client extends MatrixApi {
   }
 
   /// Presence that is set on sync.
-  PresenceType syncPresence;
+  PresenceType? syncPresence;
 
   Future<void> _checkSyncFilter() async {
     if (syncFilterId == null) {
-      syncFilterId = await defineFilter(userID, syncFilter);
+      syncFilterId = await defineFilter(userID!, syncFilter);
       await database?.storeSyncFilterId(syncFilterId, id);
     }
     return;
@@ -1090,10 +1096,10 @@ class Client extends MatrixApi {
         return;
       }
       if (database != null) {
-        _currentTransaction = database.transaction(() async {
+        _currentTransaction = database!.transaction(() async {
           await handleSync(syncResp);
           if (prevBatch != syncResp.nextBatch) {
-            await database.storePrevBatch(syncResp.nextBatch, id);
+            await database!.storePrevBatch(syncResp.nextBatch, id);
           }
         });
         await _currentTransaction;
@@ -1113,7 +1119,7 @@ class Client extends MatrixApi {
           DateTime.now().subtract(Duration(days: 30)).millisecondsSinceEpoch);
       await _updateUserDeviceKeys();
       if (encryptionEnabled) {
-        encryption.onSync();
+        encryption!.onSync();
       }
 
       // try to process the to_device queue
@@ -1146,33 +1152,33 @@ class Client extends MatrixApi {
   /// Use this method only for testing utilities!
   Future<void> handleSync(SyncUpdate sync, {bool sortAtTheEnd = false}) async {
     if (sync.toDevice != null) {
-      await _handleToDeviceEvents(sync.toDevice);
+      await _handleToDeviceEvents(sync.toDevice!);
     }
     if (sync.rooms != null) {
-      if (sync.rooms.join != null) {
-        await _handleRooms(sync.rooms.join, Membership.join,
+      if (sync.rooms!.join != null) {
+        await _handleRooms(sync.rooms!.join!, Membership.join,
             sortAtTheEnd: sortAtTheEnd);
       }
-      if (sync.rooms.invite != null) {
-        await _handleRooms(sync.rooms.invite, Membership.invite,
+      if (sync.rooms!.invite != null) {
+        await _handleRooms(sync.rooms!.invite!, Membership.invite,
             sortAtTheEnd: sortAtTheEnd);
       }
-      if (sync.rooms.leave != null) {
-        await _handleRooms(sync.rooms.leave, Membership.leave,
+      if (sync.rooms!.leave != null) {
+        await _handleRooms(sync.rooms!.leave!, Membership.leave,
             sortAtTheEnd: sortAtTheEnd);
       }
       _sortRooms();
     }
     if (sync.presence != null) {
-      for (final newPresence in sync.presence) {
+      for (final newPresence in sync.presence!) {
         presences[newPresence.senderId] = newPresence;
         onPresence.add(newPresence);
       }
     }
     if (sync.accountData != null) {
-      for (final newAccountData in sync.accountData) {
+      for (final newAccountData in sync.accountData!) {
         if (database != null) {
-          await database.storeAccountData(
+          await database!.storeAccountData(
             id,
             newAccountData.type,
             jsonEncode(newAccountData.content),
@@ -1183,10 +1189,10 @@ class Client extends MatrixApi {
       }
     }
     if (sync.deviceLists != null) {
-      await _handleDeviceListsEvents(sync.deviceLists);
+      await _handleDeviceListsEvents(sync.deviceLists!);
     }
     if (encryptionEnabled) {
-      encryption.handleDeviceOneTimeKeysCount(
+      encryption!.handleDeviceOneTimeKeysCount(
           sync.deviceOneTimeKeysCount, sync.deviceUnusedFallbackKeyTypes);
     }
     onSync.add(sync);
@@ -1194,15 +1200,15 @@ class Client extends MatrixApi {
 
   Future<void> _handleDeviceListsEvents(DeviceListsUpdate deviceLists) async {
     if (deviceLists.changed is List) {
-      for (final userId in deviceLists.changed) {
+      for (final userId in deviceLists.changed!) {
         if (_userDeviceKeys.containsKey(userId)) {
-          _userDeviceKeys[userId].outdated = true;
+          _userDeviceKeys[userId]!.outdated = true;
           if (database != null) {
-            await database.storeUserDeviceKeysInfo(id, userId, true);
+            await database!.storeUserDeviceKeysInfo(id, userId, true);
           }
         }
       }
-      for (final userId in deviceLists.left) {
+      for (final userId in deviceLists.left!) {
         if (_userDeviceKeys.containsKey(userId)) {
           _userDeviceKeys.remove(userId);
         }
@@ -1215,11 +1221,11 @@ class Client extends MatrixApi {
       var toDeviceEvent = ToDeviceEvent.fromJson(event.toJson());
       Logs().v('Got to_device event of type ${toDeviceEvent.type}');
       if (toDeviceEvent.type == EventTypes.Encrypted && encryptionEnabled) {
-        toDeviceEvent = await encryption.decryptToDeviceEvent(toDeviceEvent);
+        toDeviceEvent = await encryption!.decryptToDeviceEvent(toDeviceEvent);
         Logs().v('Decrypted type is: ${toDeviceEvent.type}');
       }
       if (encryptionEnabled) {
-        await encryption.handleToDeviceEvent(toDeviceEvent);
+        await encryption!.handleToDeviceEvent(toDeviceEvent);
       }
       onToDeviceEvent.add(toDeviceEvent);
     }
@@ -1241,11 +1247,11 @@ class Client extends MatrixApi {
       if (database != null) {
         // TODO: This method seems to be rather slow for some updates
         // Perhaps don't dynamically build that one query?
-        await database.storeRoomUpdate(this.id, update, getRoomById(id));
+        await database!.storeRoomUpdate(this.id, update, getRoomById(id));
       }
       _updateRoomsByRoomUpdate(update);
       final roomObj = getRoomById(id);
-      if (update.limitedTimeline && roomObj != null) {
+      if (update.limitedTimeline! && roomObj != null) {
         roomObj.resetSortOrder();
       }
       onRoomUpdate.add(update);
@@ -1257,14 +1263,14 @@ class Client extends MatrixApi {
         if (room.state?.isNotEmpty ?? false) {
           // TODO: This method seems to be comperatively slow for some updates
           await _handleRoomEvents(id,
-              room.state.map((i) => i.toJson()).toList(), EventUpdateType.state,
+              room.state!.map((i) => i.toJson()).toList(), EventUpdateType.state,
               sortAtTheEnd: sortAtTheEnd);
           handledEvents = true;
         }
         if (room.timeline?.events?.isNotEmpty ?? false) {
           await _handleRoomEvents(
               id,
-              room.timeline.events.map((i) => i.toJson()).toList(),
+              room.timeline!.events!.map((i) => i.toJson()).toList(),
               sortAtTheEnd ? EventUpdateType.history : EventUpdateType.timeline,
               sortAtTheEnd: sortAtTheEnd);
           handledEvents = true;
@@ -1272,12 +1278,12 @@ class Client extends MatrixApi {
         if (room.ephemeral?.isNotEmpty ?? false) {
           // TODO: This method seems to be comperatively slow for some updates
           await _handleEphemerals(
-              id, room.ephemeral.map((i) => i.toJson()).toList());
+              id, room.ephemeral!.map((i) => i.toJson()).toList());
         }
         if (room.accountData?.isNotEmpty ?? false) {
           await _handleRoomEvents(
               id,
-              room.accountData.map((i) => i.toJson()).toList(),
+              room.accountData!.map((i) => i.toJson()).toList(),
               EventUpdateType.accountData);
         }
       }
@@ -1285,20 +1291,20 @@ class Client extends MatrixApi {
         if (room.timeline?.events?.isNotEmpty ?? false) {
           await _handleRoomEvents(
               id,
-              room.timeline.events.map((i) => i.toJson()).toList(),
+              room.timeline!.events!.map((i) => i.toJson()).toList(),
               EventUpdateType.timeline);
           handledEvents = true;
         }
         if (room.accountData?.isNotEmpty ?? false) {
           await _handleRoomEvents(
               id,
-              room.accountData.map((i) => i.toJson()).toList(),
+              room.accountData!.map((i) => i.toJson()).toList(),
               EventUpdateType.accountData);
         }
         if (room.state?.isNotEmpty ?? false) {
           await _handleRoomEvents(
               id,
-              room.state.map((i) => i.toJson()).toList(),
+              room.state!.map((i) => i.toJson()).toList(),
               EventUpdateType.state);
           handledEvents = true;
         }
@@ -1307,7 +1313,7 @@ class Client extends MatrixApi {
           (room.inviteState?.isNotEmpty ?? false)) {
         await _handleRoomEvents(
             id,
-            room.inviteState.map((i) => i.toJson()).toList(),
+            room.inviteState!.map((i) => i.toJson()).toList(),
             EventUpdateType.inviteState);
       }
       if (handledEvents && database != null && roomObj != null) {
@@ -1330,7 +1336,7 @@ class Client extends MatrixApi {
         final receiptStateContent =
             room.roomAccountData['m.receipt']?.content ?? {};
         for (final eventEntry in event['content'].entries) {
-          final String eventID = eventEntry.key;
+          final String? eventID = eventEntry.key;
           if (event['content'][eventID]['m.read'] != null) {
             final Map<String, dynamic> userTimestampMap =
                 event['content'][eventID]['m.read'];
@@ -1338,7 +1344,7 @@ class Client extends MatrixApi {
               final mxid = userTimestampMapEntry.key;
 
               // Remove previous receipt event from this user
-              if (receiptStateContent[eventID] is Map<String, dynamic> &&
+              if (receiptStateContent[eventID!] is Map<String, dynamic> &&
                   receiptStateContent[eventID]['m.read']
                       is Map<String, dynamic> &&
                   receiptStateContent[eventID]['m.read'].containsKey(mxid)) {
@@ -1402,24 +1408,24 @@ class Client extends MatrixApi {
           database != null &&
           room.getState(EventTypes.RoomMember, event['sender']) == null) {
         // In order to correctly render room list previews we need to fetch the member from the database
-        final user = await database.getUser(id, event['sender'], room);
+        final user = await database!.getUser(id, event['sender'], room);
         if (user != null) {
           room.setState(user);
         }
       }
       if (type != EventUpdateType.ephemeral && database != null) {
-        await database.storeEventUpdate(id, update);
+        await database!.storeEventUpdate(id, update);
       }
       _updateRoomsByEventUpdate(update);
       if (encryptionEnabled) {
-        await encryption.handleEventUpdate(update);
+        await encryption!.handleEventUpdate(update);
       }
       onEvent.add(update);
 
       final rawUnencryptedEvent = update.content;
 
       if (prevBatch != null && type == EventUpdateType.timeline) {
-        if (rawUnencryptedEvent['type'] == EventTypes.CallInvite) {
+        if (rawUnencryptedEvent!['type'] == EventTypes.CallInvite) {
           onCallInvite
               .add(Event.fromJson(rawUnencryptedEvent, room, sortOrder));
         } else if (rawUnencryptedEvent['type'] == EventTypes.CallHangup) {
@@ -1441,9 +1447,9 @@ class Client extends MatrixApi {
     // Search the room in the rooms
     num j = 0;
     for (j = 0; j < rooms.length; j++) {
-      if (rooms[j].id == chatUpdate.id) break;
+      if (rooms[j as int]!.id == chatUpdate.id) break;
     }
-    final found = (j < rooms.length && rooms[j].id == chatUpdate.id);
+    final found = (j < rooms.length && rooms[j as int]!.id == chatUpdate.id);
     final isLeftRoom = chatUpdate.membership == Membership.leave;
 
     // Does the chat already exist in the list rooms?
@@ -1454,40 +1460,40 @@ class Client extends MatrixApi {
         id: chatUpdate.id,
         membership: chatUpdate.membership,
         prev_batch: chatUpdate.prev_batch,
-        highlightCount: chatUpdate.highlight_count,
-        notificationCount: chatUpdate.notification_count,
+        highlightCount: chatUpdate.highlight_count as int?,
+        notificationCount: chatUpdate.notification_count as int?,
         summary: chatUpdate.summary,
         roomAccountData: {},
         client: this,
       );
-      rooms.insert(position, newRoom);
+      rooms.insert(position as int, newRoom);
     }
     // If the membership is "leave" then remove the item and stop here
     else if (found && isLeftRoom) {
-      rooms.removeAt(j);
+      rooms.removeAt(j as int);
     }
     // Update notification, highlight count and/or additional informations
     else if (found &&
         chatUpdate.membership != Membership.leave &&
-        (rooms[j].membership != chatUpdate.membership ||
-            rooms[j].notificationCount != chatUpdate.notification_count ||
-            rooms[j].highlightCount != chatUpdate.highlight_count ||
+        (rooms[j as int]!.membership != chatUpdate.membership ||
+            rooms[j]!.notificationCount != chatUpdate.notification_count ||
+            rooms[j]!.highlightCount != chatUpdate.highlight_count ||
             chatUpdate.summary != null)) {
-      rooms[j].membership = chatUpdate.membership;
-      rooms[j].notificationCount = chatUpdate.notification_count;
-      rooms[j].highlightCount = chatUpdate.highlight_count;
+      rooms[j]!.membership = chatUpdate.membership;
+      rooms[j]!.notificationCount = chatUpdate.notification_count as int?;
+      rooms[j]!.highlightCount = chatUpdate.highlight_count as int?;
       if (chatUpdate.prev_batch != null) {
-        rooms[j].prev_batch = chatUpdate.prev_batch;
+        rooms[j]!.prev_batch = chatUpdate.prev_batch;
       }
       if (chatUpdate.summary != null) {
-        final roomSummaryJson = rooms[j].summary.toJson()
-          ..addAll(chatUpdate.summary.toJson());
-        rooms[j].summary = RoomSummary.fromJson(roomSummaryJson);
+        final roomSummaryJson = rooms[j]!.summary.toJson()
+          ..addAll(chatUpdate.summary!.toJson());
+        rooms[j]!.summary = RoomSummary.fromJson(roomSummaryJson);
       }
-      if (rooms[j].onUpdate != null) rooms[j].onUpdate.add(rooms[j].id);
-      if (chatUpdate.limitedTimeline && requestHistoryOnLimitedTimeline) {
-        Logs().v('Limited timeline for ${rooms[j].id} request history now');
-        runInRoot(rooms[j].requestHistory);
+      if (rooms[j]!.onUpdate != null) rooms[j]!.onUpdate.add(rooms[j]!.id);
+      if (chatUpdate.limitedTimeline! && requestHistoryOnLimitedTimeline) {
+        Logs().v('Limited timeline for ${rooms[j]!.id} request history now');
+        runInRoot(rooms[j]!.requestHistory);
       }
     }
   }
@@ -1503,11 +1509,11 @@ class Client extends MatrixApi {
       case EventUpdateType.state:
       case EventUpdateType.inviteState:
         final stateEvent =
-            Event.fromJson(eventUpdate.content, room, eventUpdate.sortOrder);
+            Event.fromJson(eventUpdate.content!, room, eventUpdate.sortOrder);
         final prevState = room.getState(stateEvent.type, stateEvent.stateKey);
         if (eventUpdate.type == EventUpdateType.timeline &&
             prevState != null &&
-            prevState.sortOrder > stateEvent.sortOrder) {
+            prevState.sortOrder! > stateEvent.sortOrder!) {
           Logs().w('''
 A new ${eventUpdate.type} event of the type ${stateEvent.type} has arrived with a previews
 sort order ${stateEvent.sortOrder} than the current ${stateEvent.type} event with a
@@ -1515,7 +1521,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
           return;
         }
         if (stateEvent.type == EventTypes.Redaction) {
-          final String redacts = eventUpdate.content['redacts'];
+          final String? redacts = eventUpdate.content!['redacts'];
           room.states.forEach(
             (String key, Map<String, Event> states) => states.forEach(
               (String key, Event state) {
@@ -1530,12 +1536,12 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
         }
         break;
       case EventUpdateType.accountData:
-        room.roomAccountData[eventUpdate.content['type']] =
-            BasicRoomEvent.fromJson(eventUpdate.content);
+        room.roomAccountData[eventUpdate.content!['type']] =
+            BasicRoomEvent.fromJson(eventUpdate.content!);
         break;
       case EventUpdateType.ephemeral:
-        room.ephemerals[eventUpdate.content['type']] =
-            BasicRoomEvent.fromJson(eventUpdate.content);
+        room.ephemerals[eventUpdate.content!['type']] =
+            BasicRoomEvent.fromJson(eventUpdate.content!);
         break;
       case EventUpdateType.history:
         break;
@@ -1551,10 +1557,10 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   /// The compare function how the rooms should be sorted internally. By default
   /// rooms are sorted by timestamp of the last m.room.message event or the last
   /// event if there is no known message.
-  RoomSorter get sortRoomsBy => (a, b) => (a.isFavourite != b.isFavourite)
+  RoomSorter get sortRoomsBy => (a, b) => (a!.isFavourite != b!.isFavourite)
       ? (a.isFavourite ? -1 : 1)
       : (pinUnreadRooms && a.notificationCount != b.notificationCount)
-          ? b.notificationCount.compareTo(a.notificationCount)
+          ? b.notificationCount!.compareTo(a.notificationCount!)
           : b.timeCreated.millisecondsSinceEpoch
               .compareTo(a.timeCreated.millisecondsSinceEpoch);
 
@@ -1566,14 +1572,14 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   }
 
   /// A map of known device keys per user.
-  Map<String, DeviceKeysList> get userDeviceKeys => _userDeviceKeys;
-  Map<String, DeviceKeysList> _userDeviceKeys = {};
+  Map<String?, DeviceKeysList> get userDeviceKeys => _userDeviceKeys;
+  Map<String?, DeviceKeysList> _userDeviceKeys = {};
 
   /// Gets user device keys by its curve25519 key. Returns null if it isn't found
-  DeviceKeys getUserDeviceKeysByCurve25519Key(String senderKey) {
+  DeviceKeys? getUserDeviceKeysByCurve25519Key(String? senderKey) {
     for (final user in userDeviceKeys.values) {
       final device = user.deviceKeys.values
-          .firstWhere((e) => e.curve25519Key == senderKey, orElse: () => null);
+          .firstWhere((e) => e!.curve25519Key == senderKey, orElse: () => null);
       if (device != null) {
         return device;
       }
@@ -1581,10 +1587,10 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     return null;
   }
 
-  Future<Set<String>> _getUserIdsInEncryptedRooms() async {
-    final userIds = <String>{};
+  Future<Set<String?>> _getUserIdsInEncryptedRooms() async {
+    final userIds = <String?>{};
     for (final room in rooms) {
-      if (room.encrypted) {
+      if (room!.encrypted) {
         try {
           final userList = await room.requestParticipants();
           for (final user in userList) {
@@ -1613,37 +1619,37 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
 
       // Remove all userIds we no longer need to track the devices of.
       _userDeviceKeys
-          .removeWhere((String userId, v) => !trackedUserIds.contains(userId));
+          .removeWhere((String? userId, v) => !trackedUserIds.contains(userId));
 
       // Check if there are outdated device key lists. Add it to the set.
-      final outdatedLists = <String, List<String>>{};
+      final outdatedLists = <String?, List<String>>{};
       for (final userId in trackedUserIds) {
         if (!userDeviceKeys.containsKey(userId)) {
           _userDeviceKeys[userId] = DeviceKeysList(userId, this);
         }
-        final deviceKeysList = userDeviceKeys[userId];
-        if (deviceKeysList.outdated &&
-            (!_keyQueryFailures.containsKey(userId.domain) ||
+        final deviceKeysList = userDeviceKeys[userId]!;
+        if (deviceKeysList.outdated! &&
+            (!_keyQueryFailures.containsKey(userId!.domain) ||
                 DateTime.now()
                     .subtract(Duration(minutes: 5))
-                    .isAfter(_keyQueryFailures[userId.domain]))) {
+                    .isAfter(_keyQueryFailures[userId.domain!]!))) {
           outdatedLists[userId] = [];
         }
       }
 
       if (outdatedLists.isNotEmpty) {
         // Request the missing device key lists from the server.
-        final response = await queryKeys(outdatedLists, timeout: 10000);
+        final response = await queryKeys(outdatedLists as Map<String, List<String>>, timeout: 10000);
         if (!isLogged()) return;
 
-        for (final rawDeviceKeyListEntry in response.deviceKeys.entries) {
+        for (final rawDeviceKeyListEntry in response.deviceKeys!.entries) {
           final userId = rawDeviceKeyListEntry.key;
           if (!userDeviceKeys.containsKey(userId)) {
             _userDeviceKeys[userId] = DeviceKeysList(userId, this);
           }
           final oldKeys =
-              Map<String, DeviceKeys>.from(_userDeviceKeys[userId].deviceKeys);
-          _userDeviceKeys[userId].deviceKeys = {};
+              Map<String, DeviceKeys>.from(_userDeviceKeys[userId]!.deviceKeys);
+          _userDeviceKeys[userId]!.deviceKeys = {};
           for (final rawDeviceKeyEntry in rawDeviceKeyListEntry.value.entries) {
             final deviceId = rawDeviceKeyEntry.key;
 
@@ -1654,35 +1660,35 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
               // is this a new key or the same one as an old one?
               // better store an update - the signatures might have changed!
               if (!oldKeys.containsKey(deviceId) ||
-                  oldKeys[deviceId].ed25519Key == entry.ed25519Key) {
+                  oldKeys[deviceId]!.ed25519Key == entry.ed25519Key) {
                 if (oldKeys.containsKey(deviceId)) {
                   // be sure to save the verified status
-                  entry.setDirectVerified(oldKeys[deviceId].directVerified);
-                  entry.blocked = oldKeys[deviceId].blocked;
-                  entry.validSignatures = oldKeys[deviceId].validSignatures;
+                  entry.setDirectVerified(oldKeys[deviceId]!.directVerified);
+                  entry.blocked = oldKeys[deviceId]!.blocked;
+                  entry.validSignatures = oldKeys[deviceId]!.validSignatures;
                 }
-                _userDeviceKeys[userId].deviceKeys[deviceId] = entry;
+                _userDeviceKeys[userId]!.deviceKeys[deviceId] = entry;
                 if (deviceId == deviceID &&
                     entry.ed25519Key == fingerprintKey) {
                   // Always trust the own device
                   entry.setDirectVerified(true);
                 }
                 if (database != null) {
-                  dbActions.add(() => database.storeUserDeviceKey(
+                  dbActions.add(() => database!.storeUserDeviceKey(
                         id,
                         userId,
                         deviceId,
                         json.encode(entry.toJson()),
                         entry.directVerified,
                         entry.blocked,
-                        entry.lastActive.millisecondsSinceEpoch,
+                        entry.lastActive!.millisecondsSinceEpoch,
                       ));
                 }
               } else if (oldKeys.containsKey(deviceId)) {
                 // This shouldn't ever happen. The same device ID has gotten
                 // a new public key. So we ignore the update. TODO: ask krille
                 // if we should instead use the new key with unknown verified / blocked status
-                _userDeviceKeys[userId].deviceKeys[deviceId] =
+                _userDeviceKeys[userId]!.deviceKeys[deviceId] =
                     oldKeys[deviceId];
               }
             } else {
@@ -1693,17 +1699,17 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
           if (database != null) {
             for (final oldDeviceKeyEntry in oldKeys.entries) {
               final deviceId = oldDeviceKeyEntry.key;
-              if (!_userDeviceKeys[userId].deviceKeys.containsKey(deviceId)) {
+              if (!_userDeviceKeys[userId]!.deviceKeys.containsKey(deviceId)) {
                 // we need to remove an old key
                 dbActions.add(
-                    () => database.removeUserDeviceKey(id, userId, deviceId));
+                    () => database!.removeUserDeviceKey(id, userId, deviceId));
               }
             }
           }
-          _userDeviceKeys[userId].outdated = false;
+          _userDeviceKeys[userId]!.outdated = false;
           if (database != null) {
             dbActions
-                .add(() => database.storeUserDeviceKeysInfo(id, userId, false));
+                .add(() => database!.storeUserDeviceKeysInfo(id, userId, false));
           }
         }
         // next we parse and persist the cross signing keys
@@ -1724,17 +1730,17 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
               _userDeviceKeys[userId] = DeviceKeysList(userId, this);
             }
             final oldKeys = Map<String, CrossSigningKey>.from(
-                _userDeviceKeys[userId].crossSigningKeys);
-            _userDeviceKeys[userId].crossSigningKeys = {};
+                _userDeviceKeys[userId]!.crossSigningKeys);
+            _userDeviceKeys[userId]!.crossSigningKeys = {};
             // add the types we aren't handling atm back
             for (final oldEntry in oldKeys.entries) {
-              if (!oldEntry.value.usage.contains(keyType)) {
-                _userDeviceKeys[userId].crossSigningKeys[oldEntry.key] =
+              if (!oldEntry.value.usage!.contains(keyType)) {
+                _userDeviceKeys[userId]!.crossSigningKeys[oldEntry.key] =
                     oldEntry.value;
               } else if (database != null) {
                 // There is a previous cross-signing key with  this usage, that we no
                 // longer need/use. Clear it from the database.
-                dbActions.add(() => database.removeUserCrossSigningKey(
+                dbActions.add(() => database!.removeUserCrossSigningKey(
                     id, userId, oldEntry.key));
               }
             }
@@ -1743,23 +1749,23 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
             if (entry.isValid) {
               final publicKey = entry.publicKey;
               if (!oldKeys.containsKey(publicKey) ||
-                  oldKeys[publicKey].ed25519Key == entry.ed25519Key) {
+                  oldKeys[publicKey!]!.ed25519Key == entry.ed25519Key) {
                 if (oldKeys.containsKey(publicKey)) {
                   // be sure to save the verification status
-                  entry.setDirectVerified(oldKeys[publicKey].directVerified);
-                  entry.blocked = oldKeys[publicKey].blocked;
-                  entry.validSignatures = oldKeys[publicKey].validSignatures;
+                  entry.setDirectVerified(oldKeys[publicKey!]!.directVerified);
+                  entry.blocked = oldKeys[publicKey]!.blocked;
+                  entry.validSignatures = oldKeys[publicKey]!.validSignatures;
                 }
-                _userDeviceKeys[userId].crossSigningKeys[publicKey] = entry;
+                _userDeviceKeys[userId]!.crossSigningKeys[publicKey] = entry;
               } else {
                 // This shouldn't ever happen. The same device ID has gotten
                 // a new public key. So we ignore the update. TODO: ask krille
                 // if we should instead use the new key with unknown verified / blocked status
-                _userDeviceKeys[userId].crossSigningKeys[publicKey] =
+                _userDeviceKeys[userId]!.crossSigningKeys[publicKey] =
                     oldKeys[publicKey];
               }
               if (database != null) {
-                dbActions.add(() => database.storeUserCrossSigningKey(
+                dbActions.add(() => database!.storeUserCrossSigningKey(
                       id,
                       userId,
                       publicKey,
@@ -1769,17 +1775,17 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
                     ));
               }
             }
-            _userDeviceKeys[userId].outdated = false;
+            _userDeviceKeys[userId]!.outdated = false;
             if (database != null) {
               dbActions.add(
-                  () => database.storeUserDeviceKeysInfo(id, userId, false));
+                  () => database!.storeUserDeviceKeysInfo(id, userId, false));
             }
           }
         }
 
         // now process all the failures
         if (response.failures != null) {
-          for (final failureDomain in response.failures.keys) {
+          for (final failureDomain in response.failures!.keys) {
             _keyQueryFailures[failureDomain] = DateTime.now();
           }
         }
@@ -1807,7 +1813,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     if (database == null || !_toDeviceQueueNeedsProcessing) {
       return;
     }
-    final entries = await database.getToDeviceEventQueue(id);
+    final entries = await database!.getToDeviceEventQueue(id);
     if (entries.isEmpty) {
       _toDeviceQueueNeedsProcessing = false;
       return;
@@ -1815,14 +1821,14 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     for (final entry in entries) {
       // Convert the Json Map to the correct format regarding
       // https: //matrix.org/docs/spec/client_server/r0.6.1#put-matrix-client-r0-sendtodevice-eventtype-txnid
-      final data = entry.content.map((k, v) =>
+      final data = entry.content!.map((k, v) =>
           MapEntry<String, Map<String, Map<String, dynamic>>>(
               k,
               (v as Map).map((k, v) => MapEntry<String, Map<String, dynamic>>(
                   k, Map<String, dynamic>.from(v)))));
 
-      await super.sendToDevice(entry.type, entry.txnId, data);
-      await database.deleteFromToDeviceQueue(id, entry.id);
+      await super.sendToDevice(entry.type!, entry.txnId!, data);
+      await database!.deleteFromToDeviceQueue(id, entry.id);
     }
   }
 
@@ -1833,11 +1839,11 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   Future<void> sendToDevice(
     String eventType,
     String txnId,
-    Map<String, Map<String, Map<String, dynamic>>> messages,
+    Map<String?, Map<String?, Map<String, dynamic>>> messages,
   ) async {
     try {
       await processToDeviceQueue();
-      await super.sendToDevice(eventType, txnId, messages);
+      await super.sendToDevice(eventType, txnId, messages as Map<String, Map<String, Map<String, dynamic>>>);
     } catch (e, s) {
       Logs().w(
           '[Client] Problem while sending to_device event, retrying later...',
@@ -1845,7 +1851,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
           s);
       if (database != null) {
         _toDeviceQueueNeedsProcessing = true;
-        await database.insertIntoToDeviceQueue(
+        await database!.insertIntoToDeviceQueue(
             id, eventType, txnId, json.encode(messages));
       }
       rethrow;
@@ -1855,38 +1861,38 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   /// Send an (unencrypted) to device [message] of a specific [eventType] to all
   /// devices of a set of [users].
   Future<void> sendToDevicesOfUserIds(
-    Set<String> users,
+    Set<String?> users,
     String eventType,
     Map<String, dynamic> message, {
-    String messageId,
+    String? messageId,
   }) async {
     // Send with send-to-device messaging
-    final data = <String, Map<String, Map<String, dynamic>>>{};
+    final data = <String?, Map<String, Map<String, dynamic>>>{};
     for (final user in users) {
       data[user] = {};
-      data[user]['*'] = message;
+      data[user]!['*'] = message;
     }
     await sendToDevice(
         eventType, messageId ?? generateUniqueTransactionId(), data);
     return;
   }
 
-  final MultiLock<DeviceKeys> _sendToDeviceEncryptedLock = MultiLock();
+  final MultiLock<DeviceKeys?> _sendToDeviceEncryptedLock = MultiLock();
 
   /// Sends an encrypted [message] of this [eventType] to these [deviceKeys].
   Future<void> sendToDeviceEncrypted(
-    List<DeviceKeys> deviceKeys,
-    String eventType,
-    Map<String, dynamic> message, {
-    String messageId,
+    List<DeviceKeys?> deviceKeys,
+    String? eventType,
+    Map<String, dynamic>? message, {
+    String? messageId,
     bool onlyVerified = false,
   }) async {
     if (!encryptionEnabled) return;
     // Don't send this message to blocked devices, and if specified onlyVerified
     // then only send it to verified devices
     if (deviceKeys.isNotEmpty) {
-      deviceKeys.removeWhere((DeviceKeys deviceKeys) =>
-          deviceKeys.blocked ||
+      deviceKeys.removeWhere((DeviceKeys? deviceKeys) =>
+          deviceKeys!.blocked ||
           (deviceKeys.userId == userID && deviceKeys.deviceId == deviceID) ||
           (onlyVerified && !deviceKeys.verified));
       if (deviceKeys.isEmpty) return;
@@ -1905,8 +1911,8 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
 
       // Send with send-to-device messaging
       var data = <String, Map<String, Map<String, dynamic>>>{};
-      data = await encryption.encryptToDeviceMessage(
-          deviceKeys, eventType, message);
+      data = await (encryption!.encryptToDeviceMessage(
+          deviceKeys, eventType, message) as FutureOr<Map<String, Map<String, Map<String, dynamic>>>>);
       eventType = EventTypes.Encrypted;
       await sendToDevice(
           eventType, messageId ?? generateUniqueTransactionId(), data);
@@ -1920,21 +1926,21 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   /// foreground. It automatically chunks sending to device keys based on
   /// activity.
   Future<void> sendToDeviceEncryptedChunked(
-    List<DeviceKeys> deviceKeys,
+    List<DeviceKeys?> deviceKeys,
     String eventType,
     Map<String, dynamic> message,
   ) async {
     if (!encryptionEnabled) return;
     // be sure to copy our device keys list
     deviceKeys = List<DeviceKeys>.from(deviceKeys);
-    deviceKeys.removeWhere((DeviceKeys k) =>
-        k.blocked || (k.userId == userID && k.deviceId == deviceID));
+    deviceKeys.removeWhere(((DeviceKeys k) =>
+        k.blocked || (k.userId == userID && k.deviceId == deviceID)) as bool Function(DeviceKeys?));
     if (deviceKeys.isEmpty) return;
     message = message.copy(); // make sure we deep-copy the message
     // make sure all the olm sessions are loaded from database
     Logs().v('Sending to device chunked... (${deviceKeys.length} devices)');
     // sort so that devices we last received messages from get our message first
-    deviceKeys.sort((keyA, keyB) => keyB.lastActive.compareTo(keyA.lastActive));
+    deviceKeys.sort(((keyA, keyB) => keyB.lastActive!.compareTo(keyA.lastActive!)) as int Function(DeviceKeys?, DeviceKeys?)?);
     // and now send out in chunks of 20
     const chunkSize = 20;
 
@@ -1948,7 +1954,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
           i,
           i + chunkSize > deviceKeys.length
               ? deviceKeys.length
-              : i + chunkSize);
+              : i + chunkSize) as List<DeviceKeys>;
       // and send
       await sendToDeviceEncrypted(chunk, eventType, message);
     }
@@ -1975,13 +1981,13 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
 
   /// Whether all push notifications are muted using the [.m.rule.master]
   /// rule of the push rules: https://matrix.org/docs/spec/client_server/r0.6.0#m-rule-master
-  bool get allPushNotificationsMuted {
+  bool? get allPushNotificationsMuted {
     if (!accountData.containsKey('m.push_rules') ||
-        !(accountData['m.push_rules'].content['global'] is Map)) {
+        !(accountData['m.push_rules']!.content['global'] is Map)) {
       return false;
     }
-    final Map<String, dynamic> globalPushRules =
-        accountData['m.push_rules'].content['global'];
+    final Map<String, dynamic>? globalPushRules =
+        accountData['m.push_rules']!.content['global'];
     if (globalPushRules == null) return false;
 
     if (globalPushRules['override'] is List) {
@@ -2007,11 +2013,11 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   /// Changes the password. You should either set oldPasswort or another authentication flow.
   @override
   Future<void> changePassword(String newPassword,
-      {String oldPassword, AuthenticationData auth, bool logoutDevices}) async {
+      {String? oldPassword, AuthenticationData? auth, bool? logoutDevices}) async {
     try {
       if (oldPassword != null) {
         auth = AuthenticationPassword(
-          identifier: AuthenticationUserIdentifier(user: userID),
+          identifier: AuthenticationUserIdentifier(user: userID!),
           password: oldPassword,
         );
       }
@@ -2021,8 +2027,8 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
       if (!matrixException.requireAdditionalAuthentication) {
         rethrow;
       }
-      if (matrixException.authenticationFlows.length != 1 ||
-          !matrixException.authenticationFlows.first.stages
+      if (matrixException.authenticationFlows!.length != 1 ||
+          !matrixException.authenticationFlows!.first.stages
               .contains(AuthenticationTypes.password)) {
         rethrow;
       }
@@ -2032,7 +2038,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
       return changePassword(
         newPassword,
         auth: AuthenticationPassword(
-          identifier: AuthenticationUserIdentifier(user: userID),
+          identifier: AuthenticationUserIdentifier(user: userID!),
           password: oldPassword,
           session: matrixException.session,
         ),
@@ -2064,9 +2070,9 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
   /// A list of mxids of users who are ignored.
   List<String> get ignoredUsers => (accountData
               .containsKey('m.ignored_user_list') &&
-          accountData['m.ignored_user_list'].content['ignored_users'] is Map)
+          accountData['m.ignored_user_list']!.content['ignored_users'] is Map)
       ? List<String>.from(
-          accountData['m.ignored_user_list'].content['ignored_users'].keys)
+          accountData['m.ignored_user_list']!.content['ignored_users'].keys)
       : [];
 
   /// Ignore another user. This will clear the local cached messages to
@@ -2075,7 +2081,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     if (!userId.isValidMatrixId) {
       throw Exception('$userId is not a valid mxid!');
     }
-    await setAccountData(userID, 'm.ignored_user_list', {
+    await setAccountData(userID!, 'm.ignored_user_list', {
       'ignored_users': Map.fromEntries(
           (ignoredUsers..add(userId)).map((key) => MapEntry(key, {}))),
     });
@@ -2092,7 +2098,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     if (!ignoredUsers.contains(userId)) {
       throw Exception('$userId is not in the ignore list!');
     }
-    await setAccountData(userID, 'm.ignored_user_list', {
+    await setAccountData(userID!, 'm.ignored_user_list', {
       'ignored_users': Map.fromEntries(
           (ignoredUsers..remove(userId)).map((key) => MapEntry(key, {}))),
     });
@@ -2129,7 +2135,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     encryption = null;
     try {
       if (closeDatabase && database != null) {
-        await database
+        await database!
             .close()
             .catchError((e, s) => Logs().w('Failed to close database: ', e, s));
         _database = null;
@@ -2140,15 +2146,15 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
     return;
   }
 
-  Future<void> _migrateFromLegacyDatabase() async {
+  Future<void>? _migrateFromLegacyDatabase() async {
     Logs().i('Check legacy database for migration data...');
-    final legacyDatabase = await legacyDatabaseBuilder(this);
+    final legacyDatabase = await legacyDatabaseBuilder!(this);
     final migrateClient = await legacyDatabase.getClient(clientName);
 
     if (migrateClient != null) {
       Logs().i('Found data in the legacy database!');
       _id = migrateClient['client_id'];
-      await database.insertClient(
+      await database!.insertClient(
         clientName,
         migrateClient['homeserver_url'],
         migrateClient['token'],
@@ -2163,7 +2169,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
         final ssssCache = await legacyDatabase.getSSSSCache(_id, type);
         if (ssssCache != null) {
           Logs().d('Migrate $type...');
-          await database.storeSSSSCache(
+          await database!.storeSSSSCache(
             _id,
             type,
             ssssCache.keyId,
@@ -2176,11 +2182,11 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
       final userDeviceKeys = await legacyDatabase.getUserDeviceKeys(this);
       for (final userId in userDeviceKeys.keys) {
         Logs().d('Migrate Device Keys of user $userId...');
-        final deviceKeysList = userDeviceKeys[userId];
+        final deviceKeysList = userDeviceKeys[userId]!;
         for (final crossSigningKey in deviceKeysList.crossSigningKeys.values) {
           Logs().d(
-              'Migrate cross signing key with usage ${crossSigningKey.usage} and verified ${crossSigningKey.directVerified}...');
-          await database.storeUserCrossSigningKey(
+              'Migrate cross signing key with usage ${crossSigningKey!.usage} and verified ${crossSigningKey.directVerified}...');
+          await database!.storeUserCrossSigningKey(
             _id,
             userId,
             crossSigningKey.publicKey,
@@ -2190,19 +2196,19 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
           );
         }
         for (final deviceKeys in deviceKeysList.deviceKeys.values) {
-          Logs().d('Migrate device keys for ${deviceKeys.deviceId}...');
-          await database.storeUserDeviceKey(
+          Logs().d('Migrate device keys for ${deviceKeys!.deviceId}...');
+          await database!.storeUserDeviceKey(
             _id,
             userId,
             deviceKeys.deviceId,
             jsonEncode(deviceKeys.toJson()),
             deviceKeys.directVerified,
             deviceKeys.blocked,
-            deviceKeys.lastActive.millisecondsSinceEpoch,
+            deviceKeys.lastActive!.millisecondsSinceEpoch,
           );
         }
         Logs().d('Migrate user device keys info...');
-        await database.storeUserDeviceKeysInfo(
+        await database!.storeUserDeviceKeysInfo(
             _id, userId, deviceKeysList.outdated);
       }
       Logs().d('Migrate inbound group sessions...');
@@ -2211,7 +2217,7 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
         for (var i = 0; i < sessions.length; i++) {
           Logs().d('$i / ${sessions.length}');
           final session = sessions[i];
-          await database.storeInboundGroupSession(
+          await database!.storeInboundGroupSession(
             _id,
             session.roomId,
             session.sessionId,
@@ -2240,15 +2246,15 @@ sort order of ${prevState.sortOrder}. This should never happen...''');
 
 class SdkError {
   dynamic exception;
-  StackTrace stackTrace;
+  StackTrace? stackTrace;
 
   SdkError({this.exception, this.stackTrace});
 }
 
 class SyncStatusUpdate {
   final SyncStatus status;
-  final SdkError error;
-  final double progress;
+  final SdkError? error;
+  final double? progress;
   const SyncStatusUpdate(this.status, {this.error, this.progress});
 }
 
@@ -2270,7 +2276,7 @@ class BadServerVersionsException implements Exception {
 }
 
 class BadServerLoginTypesException implements Exception {
-  final Set<String> serverLoginTypes, supportedLoginTypes;
+  final Set<String?>? serverLoginTypes, supportedLoginTypes;
   BadServerLoginTypesException(this.serverLoginTypes, this.supportedLoginTypes);
 
   @override
