@@ -28,31 +28,31 @@ import 'utils/room_update.dart';
 /// automatically. The initial
 /// event list will be retreived when created by the `room.getTimeline()` method.
 class Timeline {
-  final Room room;
+  final Room? room;
   final List<Event> events;
 
   /// Map of event ID to map of type to set of aggregated events
-  final Map<String, Map<String, Set<Event>>> aggregatedEvents = {};
+  final Map<String?, Map<String?, Set<Event>>> aggregatedEvents = {};
 
-  final void Function() onUpdate;
-  final void Function(int insertID) onInsert;
+  final void Function()? onUpdate;
+  final void Function(int insertID)? onInsert;
 
-  StreamSubscription<EventUpdate> sub;
-  StreamSubscription<RoomUpdate> roomSub;
-  StreamSubscription<String> sessionIdReceivedSub;
+  StreamSubscription<EventUpdate>? sub;
+  StreamSubscription<RoomUpdate>? roomSub;
+  StreamSubscription<String?>? sessionIdReceivedSub;
   bool isRequestingHistory = false;
 
-  final Map<String, Event> _eventCache = {};
+  final Map<String?, Event> _eventCache = {};
 
   /// Searches for the event in this timeline. If not
   /// found, requests from the server. Requested events
   /// are cached.
-  Future<Event> getEventById(String id) async {
+  Future<Event?> getEventById(String? id) async {
     for (final event in events) {
       if (event.eventId == id) return event;
     }
     if (_eventCache.containsKey(id)) return _eventCache[id];
-    final requestedEvent = await room.getEventById(id);
+    final requestedEvent = await room!.getEventById(id!);
     if (requestedEvent == null) return null;
     _eventCache[id] = requestedEvent;
     return _eventCache[id];
@@ -74,7 +74,7 @@ class Timeline {
       {int historyCount = Room.defaultHistoryCount}) async {
     if (!isRequestingHistory) {
       isRequestingHistory = true;
-      await room.requestHistory(
+      await room!.requestHistory(
         historyCount: historyCount,
         onHistoryReceived: () {
           _collectHistoryUpdates = true;
@@ -98,24 +98,24 @@ class Timeline {
     }
     _historyUpdates.clear();
     _sort();
-    if (onUpdate != null) onUpdate();
+    if (onUpdate != null) onUpdate!();
   }
 
-  Timeline({this.room, List<Event> events, this.onUpdate, this.onInsert})
+  Timeline({this.room, required List<Event> events, this.onUpdate, this.onInsert})
       : events = events ?? [] {
-    sub ??= room.client.onEvent.stream.listen(_handleEventUpdate);
+    sub ??= room!.client!.onEvent.stream.listen(_handleEventUpdate);
     // if the timeline is limited we want to clear our events cache
     // as r.limitedTimeline can be "null" sometimes, we need to check for == true
     // as after receiving a limited timeline room update new events are expected
     // to be received via the onEvent stream, it is unneeded to call sortAndUpdate
-    roomSub ??= room.client.onRoomUpdate.stream
-        .where((r) => r.id == room.id && r.limitedTimeline == true)
+    roomSub ??= room!.client!.onRoomUpdate.stream
+        .where((r) => r.id == room!.id && r.limitedTimeline == true)
         .listen((r) {
       events.clear();
       aggregatedEvents.clear();
     });
     sessionIdReceivedSub ??=
-        room.onSessionKeyReceived.stream.listen(_sessionKeyReceived);
+        room!.onSessionKeyReceived.stream.listen(_sessionKeyReceived);
 
     // we want to populate our aggregated events
     for (final e in events) {
@@ -131,30 +131,30 @@ class Timeline {
     sessionIdReceivedSub?.cancel();
   }
 
-  void _sessionKeyReceived(String sessionId) async {
+  void _sessionKeyReceived(String? sessionId) async {
     var decryptAtLeastOneEvent = false;
     final decryptFn = () async {
-      if (!room.client.encryptionEnabled) {
+      if (!room!.client!.encryptionEnabled) {
         return;
       }
       for (var i = 0; i < events.length; i++) {
         if (events[i].type == EventTypes.Encrypted &&
             events[i].messageType == MessageTypes.BadEncrypted &&
             events[i].content['session_id'] == sessionId) {
-          events[i] = await room.client.encryption
-              .decryptRoomEvent(room.id, events[i], store: true);
+          events[i] = await room!.client!.encryption!
+              .decryptRoomEvent(room!.id, events[i], store: true);
           if (events[i].type != EventTypes.Encrypted) {
             decryptAtLeastOneEvent = true;
           }
         }
       }
     };
-    if (room.client.database != null) {
-      await room.client.database.transaction(decryptFn);
+    if (room!.client!.database != null) {
+      await room!.client!.database!.transaction(decryptFn);
     } else {
       await decryptFn();
     }
-    if (decryptAtLeastOneEvent) onUpdate();
+    if (decryptAtLeastOneEvent) onUpdate!();
   }
 
   /// Request the keys for undecryptable events of this timeline
@@ -164,7 +164,7 @@ class Timeline {
           event.messageType == MessageTypes.BadEncrypted &&
           event.content['can_request_session'] == true) {
         try {
-          room.client.encryption.keyManager.maybeAutoRequest(room.id,
+          room!.client!.encryption!.keyManager!.maybeAutoRequest(room!.id,
               event.content['session_id'], event.content['sender_key']);
         } catch (_) {
           // dispose
@@ -173,7 +173,7 @@ class Timeline {
     }
   }
 
-  int _findEvent({String event_id, String unsigned_txid}) {
+  int _findEvent({String? event_id, String? unsigned_txid}) {
     // we want to find any existing event where either the passed event_id or the passed unsigned_txid
     // matches either the event_id or transaction_id of the existing event.
     // For that we create two sets, searchNeedle, what we search, and searchHaystack, where we check if there is a match.
@@ -188,13 +188,13 @@ class Timeline {
     }
     int i;
     for (i = 0; i < events.length; i++) {
-      final searchHaystack = <String>{};
+      final searchHaystack = <String?>{};
       if (events[i].eventId != null) {
         searchHaystack.add(events[i].eventId);
       }
       if (events[i].unsigned != null &&
-          events[i].unsigned['transaction_id'] != null) {
-        searchHaystack.add(events[i].unsigned['transaction_id']);
+          events[i].unsigned!['transaction_id'] != null) {
+        searchHaystack.add(events[i].unsigned!['transaction_id']);
       }
       if (searchNeedle.intersection(searchHaystack).isNotEmpty) {
         break;
@@ -207,7 +207,7 @@ class Timeline {
     eventSet.removeWhere((e) =>
         e.matchesEventOrTransactionId(event.eventId) ||
         (event.unsigned != null &&
-            e.matchesEventOrTransactionId(event.unsigned['transaction_id'])));
+            e.matchesEventOrTransactionId(event.unsigned!['transaction_id'])));
   }
 
   void addAggregatedEvent(Event event) {
@@ -216,26 +216,26 @@ class Timeline {
       return; // nothing to do
     }
     if (!aggregatedEvents.containsKey(event.relationshipEventId)) {
-      aggregatedEvents[event.relationshipEventId] = <String, Set<Event>>{};
+      aggregatedEvents[event.relationshipEventId] = <String?, Set<Event>>{};
     }
-    if (!aggregatedEvents[event.relationshipEventId]
+    if (!aggregatedEvents[event.relationshipEventId]!
         .containsKey(event.relationshipType)) {
-      aggregatedEvents[event.relationshipEventId]
+      aggregatedEvents[event.relationshipEventId]!
           [event.relationshipType] = <Event>{};
     }
     // remove a potential old event
     _removeEventFromSet(
-        aggregatedEvents[event.relationshipEventId][event.relationshipType],
+        aggregatedEvents[event.relationshipEventId]![event.relationshipType]!,
         event);
     // add the new one
-    aggregatedEvents[event.relationshipEventId][event.relationshipType]
+    aggregatedEvents[event.relationshipEventId]![event.relationshipType]!
         .add(event);
   }
 
   void removeAggregatedEvent(Event event) {
     aggregatedEvents.remove(event.eventId);
     if (event.unsigned != null) {
-      aggregatedEvents.remove(event.unsigned['transaction_id']);
+      aggregatedEvents.remove(event.unsigned!['transaction_id']);
     }
     for (final types in aggregatedEvents.values) {
       for (final events in types.values) {
@@ -247,7 +247,7 @@ class Timeline {
   void _handleEventUpdate(EventUpdate eventUpdate,
       {bool sortAndUpdate = true}) {
     try {
-      if (eventUpdate.roomID != room.id) return;
+      if (eventUpdate.roomID != room!.id) return;
 
       if (eventUpdate.type == EventUpdateType.history &&
           _collectHistoryUpdates) {
@@ -259,37 +259,37 @@ class Timeline {
           eventUpdate.type != EventUpdateType.history) {
         return;
       }
-      final status = eventUpdate.content['status'] ??
-          (eventUpdate.content['unsigned'] is Map<String, dynamic>
-              ? eventUpdate.content['unsigned'][messageSendingStatusKey]
+      final status = eventUpdate.content!['status'] ??
+          (eventUpdate.content!['unsigned'] is Map<String, dynamic>
+              ? eventUpdate.content!['unsigned'][messageSendingStatusKey]
               : null) ??
           2;
       // Redaction events are handled as modification for existing events.
-      if (eventUpdate.content['type'] == EventTypes.Redaction) {
-        final eventId = _findEvent(event_id: eventUpdate.content['redacts']);
+      if (eventUpdate.content!['type'] == EventTypes.Redaction) {
+        final eventId = _findEvent(event_id: eventUpdate.content!['redacts']);
         if (eventId < events.length) {
           removeAggregatedEvent(events[eventId]);
           events[eventId].setRedactionEvent(
-              Event.fromJson(eventUpdate.content, room, eventUpdate.sortOrder));
+              Event.fromJson(eventUpdate.content!, room, eventUpdate.sortOrder));
         }
       } else if (status == -2) {
-        final i = _findEvent(event_id: eventUpdate.content['event_id']);
+        final i = _findEvent(event_id: eventUpdate.content!['event_id']);
         if (i < events.length) {
           removeAggregatedEvent(events[i]);
           events.removeAt(i);
         }
       } else {
         final i = _findEvent(
-            event_id: eventUpdate.content['event_id'],
-            unsigned_txid: eventUpdate.content['unsigned'] is Map
-                ? eventUpdate.content['unsigned']['transaction_id']
+            event_id: eventUpdate.content!['event_id'],
+            unsigned_txid: eventUpdate.content!['unsigned'] is Map
+                ? eventUpdate.content!['unsigned']['transaction_id']
                 : null);
 
         if (i < events.length) {
           // if the old status is larger than the new one, we also want to preserve the old status
           final oldStatus = events[i].status;
           events[i] = Event.fromJson(
-              eventUpdate.content,
+              eventUpdate.content!,
               room,
               eventUpdate.type == EventUpdateType.history
                   ? events[i].sortOrder
@@ -301,21 +301,21 @@ class Timeline {
           addAggregatedEvent(events[i]);
         } else {
           final newEvent =
-              Event.fromJson(eventUpdate.content, room, eventUpdate.sortOrder);
+              Event.fromJson(eventUpdate.content!, room, eventUpdate.sortOrder);
 
           if (eventUpdate.type == EventUpdateType.history &&
               events.indexWhere(
-                      (e) => e.eventId == eventUpdate.content['event_id']) !=
+                      (e) => e.eventId == eventUpdate.content!['event_id']) !=
                   -1) return;
 
           events.insert(0, newEvent);
           addAggregatedEvent(newEvent);
-          if (onInsert != null) onInsert(0);
+          if (onInsert != null) onInsert!(0);
         }
       }
       if (sortAndUpdate) {
         _sort();
-        if (onUpdate != null) onUpdate();
+        if (onUpdate != null) onUpdate!();
       }
     } catch (e, s) {
       Logs().w('Handle event update failed', e, s);
@@ -334,7 +334,7 @@ class Timeline {
       if (a.status == -1 && b.status != -1) {
         return -1;
       }
-      return b.sortOrder - a.sortOrder > 0 ? 1 : -1;
+      return b.sortOrder! - a.sortOrder! > 0 ? 1 : -1;
     });
     _sortLock = false;
   }
