@@ -24,14 +24,14 @@ extension CommandsClientExtension on Client {
   /// Add a command to the command handler. `command` is its name, and `callback` is the
   /// callback to invoke
   void addCommand(
-      String command, FutureOr<String> Function(CommandArgs) callback) {
+      String command, FutureOr<String?> Function(CommandArgs) callback) {
     commands[command.toLowerCase()] = callback;
   }
 
   /// Parse and execute a string, `msg` is the input. Optionally `inReplyTo` is the event being
   /// replied to and `editEventId` is the eventId of the event being replied to
-  Future<String> parseAndRunCommand(Room room, String msg,
-      {Event inReplyTo, String editEventId, String txid}) async {
+  Future<String?> parseAndRunCommand(Room room, String msg,
+      {Event? inReplyTo, String? editEventId, String? txid}) async {
     final args = CommandArgs(
       inReplyTo: inReplyTo,
       editEventId: editEventId,
@@ -42,7 +42,7 @@ extension CommandsClientExtension on Client {
     if (!msg.startsWith('/')) {
       if (commands.containsKey('send')) {
         args.msg = msg;
-        return await commands['send'](args);
+        return await commands['send']!(args);
       }
       return null;
     }
@@ -57,12 +57,12 @@ extension CommandsClientExtension on Client {
       command = msg.toLowerCase();
     }
     if (commands.containsKey(command)) {
-      return await commands[command](args);
+      return await commands[command]!(args);
     }
     if (msg.startsWith('/') && commands.containsKey('send')) {
       // re-set to include the "command"
       args.msg = msg;
-      return await commands['send'](args);
+      return await commands['send']!(args);
     }
     return null;
   }
@@ -75,7 +75,7 @@ extension CommandsClientExtension on Client {
   /// Register all default commands
   void registerDefaultCommands() {
     addCommand('send', (CommandArgs args) async {
-      return await args.room.sendTextEvent(
+      return await args.room!.sendTextEvent(
         args.msg,
         inReplyTo: args.inReplyTo,
         editEventId: args.editEventId,
@@ -84,7 +84,7 @@ extension CommandsClientExtension on Client {
       );
     });
     addCommand('me', (CommandArgs args) async {
-      return await args.room.sendTextEvent(
+      return await args.room!.sendTextEvent(
         args.msg,
         inReplyTo: args.inReplyTo,
         editEventId: args.editEventId,
@@ -94,7 +94,7 @@ extension CommandsClientExtension on Client {
       );
     });
     addCommand('plain', (CommandArgs args) async {
-      return await args.room.sendTextEvent(
+      return await args.room!.sendTextEvent(
         args.msg,
         inReplyTo: args.inReplyTo,
         editEventId: args.editEventId,
@@ -110,7 +110,7 @@ extension CommandsClientExtension on Client {
         'format': 'org.matrix.custom.html',
         'formatted_body': args.msg,
       };
-      return await args.room.sendEvent(
+      return await args.room!.sendEvent(
         event,
         inReplyTo: args.inReplyTo,
         editEventId: args.editEventId,
@@ -121,71 +121,71 @@ extension CommandsClientExtension on Client {
       if (args.inReplyTo == null) {
         return null;
       }
-      return await args.room.sendReaction(args.inReplyTo.eventId, args.msg);
+      return await args.room!.sendReaction(args.inReplyTo!.eventId, args.msg);
     });
     addCommand('join', (CommandArgs args) async {
-      await args.room.client.joinRoom(args.msg);
+      await args.room!.client!.joinRoom(args.msg!);
       return null;
     });
     addCommand('leave', (CommandArgs args) async {
-      await args.room.leave();
+      await args.room!.leave();
       return '';
     });
     addCommand('op', (CommandArgs args) async {
-      final parts = args.msg.split(' ');
+      final parts = args.msg!.split(' ');
       if (parts.isEmpty) {
         return null;
       }
-      var pl = 50;
+      int? pl = 50;
       if (parts.length >= 2) {
         pl = int.tryParse(parts[1]);
       }
       final mxid = parts.first;
-      return await args.room.setPower(mxid, pl);
+      return await args.room!.setPower(mxid, pl);
     });
     addCommand('kick', (CommandArgs args) async {
-      final parts = args.msg.split(' ');
-      await args.room.kick(parts.first);
+      final parts = args.msg!.split(' ');
+      await args.room!.kick(parts.first);
       return '';
     });
     addCommand('ban', (CommandArgs args) async {
-      final parts = args.msg.split(' ');
-      await args.room.ban(parts.first);
+      final parts = args.msg!.split(' ');
+      await args.room!.ban(parts.first);
       return '';
     });
     addCommand('unban', (CommandArgs args) async {
-      final parts = args.msg.split(' ');
-      await args.room.unban(parts.first);
+      final parts = args.msg!.split(' ');
+      await args.room!.unban(parts.first);
       return '';
     });
     addCommand('invite', (CommandArgs args) async {
-      final parts = args.msg.split(' ');
-      await args.room.invite(parts.first);
+      final parts = args.msg!.split(' ');
+      await args.room!.invite(parts.first);
       return '';
     });
     addCommand('myroomnick', (CommandArgs args) async {
-      final currentEventJson = args.room
-          .getState(EventTypes.RoomMember, args.room.client.userID)
+      final currentEventJson = args.room!
+          .getState(EventTypes.RoomMember, args.room!.client!.userID)!
           .content
           .copy();
       currentEventJson['displayname'] = args.msg;
-      return await args.room.client.setRoomStateWithKey(
-        args.room.id,
+      return await args.room!.client!.setRoomStateWithKey(
+        args.room!.id!,
         EventTypes.RoomMember,
-        args.room.client.userID,
+        args.room!.client!.userID!,
         currentEventJson,
       );
     });
     addCommand('myroomavatar', (CommandArgs args) async {
-      final currentEventJson = args.room
-          .getState(EventTypes.RoomMember, args.room.client.userID)
+      final currentEventJson = args.room!
+          .getState(EventTypes.RoomMember, args.room!.client!.userID)!
           .content
           .copy();
       currentEventJson['avatar_url'] = args.msg;
-      return await args.room.client.setRoomStateWithKey(
-        args.room.id,
+      return await args.room!.client!.setRoomStateWithKey(
+        args.room!.id!,
         EventTypes.RoomMember,
-        args.room.client.userID,
+        args.room!.client!.userID!,
         currentEventJson,
       );
     });
@@ -193,11 +193,11 @@ extension CommandsClientExtension on Client {
 }
 
 class CommandArgs {
-  String msg;
-  String editEventId;
-  Event inReplyTo;
-  Room room;
-  String txid;
+  String? msg;
+  String? editEventId;
+  Event? inReplyTo;
+  Room? room;
+  String? txid;
   CommandArgs(
       {this.msg, this.editEventId, this.inReplyTo, this.room, this.txid});
 }
