@@ -31,7 +31,7 @@ void main() {
     Logs().level = Level.error;
     var olmEnabled = true;
 
-    Client client;
+    late Client client;
 
     test('setupClient', () async {
       try {
@@ -52,7 +52,7 @@ void main() {
       final payload = <String, dynamic>{
         'fox': 'floof',
       };
-      final signedPayload = client.encryption.olmManager.signJson(payload);
+      final signedPayload = client.encryption!.olmManager.signJson(payload);
       expect(
           signedPayload.checkJsonSignature(
               client.fingerprintKey, client.userID, client.deviceID),
@@ -63,26 +63,26 @@ void main() {
       if (!olmEnabled) return;
       FakeMatrixApi.calledEndpoints.clear();
       final res =
-          await client.encryption.olmManager.uploadKeys(uploadDeviceKeys: true);
+          await client.encryption!.olmManager.uploadKeys(uploadDeviceKeys: true);
       expect(res, true);
       var sent = json.decode(
-          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload'].first);
+          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload']!.first);
       expect(sent['device_keys'] != null, true);
       expect(sent['one_time_keys'] != null, true);
       expect(sent['one_time_keys'].keys.length, 66);
       expect(sent['fallback_keys'] != null, true);
       expect(sent['fallback_keys'].keys.length, 1);
       FakeMatrixApi.calledEndpoints.clear();
-      await client.encryption.olmManager.uploadKeys();
+      await client.encryption!.olmManager.uploadKeys();
       sent = json.decode(
-          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload'].first);
+          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload']!.first);
       expect(sent['device_keys'] != null, false);
       expect(sent['fallback_keys'].keys.length, 1);
       FakeMatrixApi.calledEndpoints.clear();
-      await client.encryption.olmManager
+      await client.encryption!.olmManager
           .uploadKeys(oldKeyCount: 20, unusedFallbackKey: true);
       sent = json.decode(
-          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload'].first);
+          FakeMatrixApi.calledEndpoints['/client/r0/keys/upload']!.first);
       expect(sent['one_time_keys'].keys.length, 46);
       expect(sent['fallback_keys'].keys.length, 0);
     });
@@ -90,7 +90,7 @@ void main() {
     test('handleDeviceOneTimeKeysCount', () async {
       if (!olmEnabled) return;
       FakeMatrixApi.calledEndpoints.clear();
-      client.encryption.olmManager
+      client.encryption!.olmManager
           .handleDeviceOneTimeKeysCount({'signed_curve25519': 20}, null);
       await Future.delayed(Duration(milliseconds: 50));
       expect(
@@ -98,7 +98,7 @@ void main() {
           true);
 
       FakeMatrixApi.calledEndpoints.clear();
-      client.encryption.olmManager
+      client.encryption!.olmManager
           .handleDeviceOneTimeKeysCount({'signed_curve25519': 70}, null);
       await Future.delayed(Duration(milliseconds: 50));
       expect(
@@ -106,7 +106,7 @@ void main() {
           false);
 
       FakeMatrixApi.calledEndpoints.clear();
-      client.encryption.olmManager.handleDeviceOneTimeKeysCount(null, []);
+      client.encryption!.olmManager.handleDeviceOneTimeKeysCount(null, []);
       await Future.delayed(Duration(milliseconds: 50));
       expect(
           FakeMatrixApi.calledEndpoints.containsKey('/client/r0/keys/upload'),
@@ -114,7 +114,7 @@ void main() {
 
       // this will upload keys because we assume the key count is 0, if the server doesn't send one
       FakeMatrixApi.calledEndpoints.clear();
-      client.encryption.olmManager
+      client.encryption!.olmManager
           .handleDeviceOneTimeKeysCount(null, ['signed_curve25519']);
       await Future.delayed(Duration(milliseconds: 50));
       expect(
@@ -124,30 +124,30 @@ void main() {
 
     test('restoreOlmSession', () async {
       if (!olmEnabled) return;
-      client.encryption.olmManager.olmSessions.clear();
-      await client.encryption.olmManager
+      client.encryption!.olmManager.olmSessions.clear();
+      await client.encryption!.olmManager
           .restoreOlmSession(client.userID, client.identityKey);
-      expect(client.encryption.olmManager.olmSessions.length, 1);
+      expect(client.encryption!.olmManager.olmSessions.length, 1);
 
-      client.encryption.olmManager.olmSessions.clear();
-      await client.encryption.olmManager
+      client.encryption!.olmManager.olmSessions.clear();
+      await client.encryption!.olmManager
           .restoreOlmSession(client.userID, 'invalid');
-      expect(client.encryption.olmManager.olmSessions.length, 0);
+      expect(client.encryption!.olmManager.olmSessions.length, 0);
 
-      client.encryption.olmManager.olmSessions.clear();
-      await client.encryption.olmManager
+      client.encryption!.olmManager.olmSessions.clear();
+      await client.encryption!.olmManager
           .restoreOlmSession('invalid', client.identityKey);
-      expect(client.encryption.olmManager.olmSessions.length, 0);
+      expect(client.encryption!.olmManager.olmSessions.length, 0);
     });
 
     test('startOutgoingOlmSessions', () async {
       if (!olmEnabled) return;
       // start an olm session.....with ourself!
-      client.encryption.olmManager.olmSessions.clear();
-      await client.encryption.olmManager.startOutgoingOlmSessions(
-          [client.userDeviceKeys[client.userID].deviceKeys[client.deviceID]]);
+      client.encryption!.olmManager.olmSessions.clear();
+      await client.encryption!.olmManager.startOutgoingOlmSessions(
+          [client.userDeviceKeys[client.userID]!.deviceKeys[client.deviceID]]);
       expect(
-          client.encryption.olmManager.olmSessions
+          client.encryption!.olmManager.olmSessions
               .containsKey(client.identityKey),
           true);
     });
@@ -158,7 +158,7 @@ void main() {
       final deviceId = 'JLAFKJWSCS';
       final senderKey = 'L+4+JCl8MD63dgo8z5Ta+9QAHXiANyOVSfgbHA5d3H8';
       FakeMatrixApi.calledEndpoints.clear();
-      await client.database.setLastSentMessageUserDeviceKey(
+      await client.database!.setLastSentMessageUserDeviceKey(
           json.encode({
             'type': 'm.foxies',
             'content': {
@@ -176,7 +176,7 @@ void main() {
           'sender_key': senderKey,
         },
       );
-      await client.encryption.olmManager.handleToDeviceEvent(event);
+      await client.encryption!.olmManager.handleToDeviceEvent(event);
       expect(
           FakeMatrixApi.calledEndpoints.keys.any(
               (k) => k.startsWith('/client/r0/sendToDevice/m.room.encrypted')),
@@ -186,7 +186,7 @@ void main() {
 
       // not encrypted
       FakeMatrixApi.calledEndpoints.clear();
-      await client.database.setLastSentMessageUserDeviceKey(
+      await client.database!.setLastSentMessageUserDeviceKey(
           json.encode({
             'type': 'm.foxies',
             'content': {
@@ -202,7 +202,7 @@ void main() {
         content: {},
         encryptedContent: null,
       );
-      await client.encryption.olmManager.handleToDeviceEvent(event);
+      await client.encryption!.olmManager.handleToDeviceEvent(event);
       expect(
           FakeMatrixApi.calledEndpoints.keys.any(
               (k) => k.startsWith('/client/r0/sendToDevice/m.room.encrypted')),
@@ -210,7 +210,7 @@ void main() {
 
       // device not found
       FakeMatrixApi.calledEndpoints.clear();
-      await client.database.setLastSentMessageUserDeviceKey(
+      await client.database!.setLastSentMessageUserDeviceKey(
           json.encode({
             'type': 'm.foxies',
             'content': {
@@ -228,7 +228,7 @@ void main() {
           'sender_key': 'invalid',
         },
       );
-      await client.encryption.olmManager.handleToDeviceEvent(event);
+      await client.encryption!.olmManager.handleToDeviceEvent(event);
       expect(
           FakeMatrixApi.calledEndpoints.keys.any(
               (k) => k.startsWith('/client/r0/sendToDevice/m.room.encrypted')),
