@@ -23,19 +23,19 @@ import '../../matrix.dart';
 
 extension JsonSignatureCheckExtension on Map<String, dynamic> {
   /// Checks the signature of a signed json object.
-  bool checkJsonSignature(String key, String userId, String deviceId) {
-    final Map<String, dynamic> signatures = this['signatures'];
+  bool checkJsonSignature(String? key, String? userId, String? deviceId) {
+    final Map<String, dynamic>? signatures = this['signatures'];
     if (signatures == null || !signatures.containsKey(userId)) return false;
     remove('unsigned');
     remove('signatures');
-    if (!signatures[userId].containsKey('ed25519:$deviceId')) return false;
+    if (!signatures[userId!].containsKey('ed25519:$deviceId')) return false;
     final String signature = signatures[userId]['ed25519:$deviceId'];
     final canonical = canonicalJson.encode(this);
     final message = String.fromCharCodes(canonical);
     var isValid = false;
     final olmutil = olm.Utility();
     try {
-      olmutil.ed25519_verify(key, message, signature);
+      olmutil.ed25519_verify(key!, message, signature);
       isValid = true;
     } catch (e, s) {
       isValid = false;
