@@ -23,10 +23,10 @@ import '../room.dart';
 
 extension ImagePackRoomExtension on Room {
   /// Get all the active image packs for the specified [usage], mapped by their slug
-  Map<String, ImagePackContent> getImagePacks([ImagePackUsage usage]) {
+  Map<String, ImagePackContent> getImagePacks([ImagePackUsage? usage]) {
     final allMxcs = <Uri>{}; // used for easy deduplication
     final packs = <String, ImagePackContent>{};
-    final addImagePack = (BasicEvent event, {Room room, String slug}) {
+    final Null Function(BasicEvent?, {Room room, String? slug}) addImagePack = (BasicEvent? event, {Room? room, String? slug}) {
       if (event == null) return;
       final imagePack = event.parsedImagePackContent;
       final finalSlug = slugify(slug ?? 'pack');
@@ -43,26 +43,26 @@ extension ImagePackRoomExtension on Room {
         }
         if (!packs.containsKey(finalSlug)) {
           packs[finalSlug] = ImagePackContent.fromJson(<String, dynamic>{});
-          packs[finalSlug].pack.displayName = imagePack.pack.displayName ??
+          packs[finalSlug]!.pack.displayName = imagePack.pack.displayName ??
               room?.displayname ??
               finalSlug ??
               '';
-          packs[finalSlug].pack.avatarUrl =
+          packs[finalSlug]!.pack.avatarUrl =
               imagePack.pack.avatarUrl ?? room?.avatar;
-          packs[finalSlug].pack.attribution = imagePack.pack.attribution;
+          packs[finalSlug]!.pack.attribution = imagePack.pack.attribution;
         }
-        packs[finalSlug].images[entry.key] = image;
+        packs[finalSlug]!.images[entry.key] = image;
         allMxcs.add(image.url);
       }
     };
     // first we add the user image pack
-    addImagePack(client.accountData['im.ponies.user_emotes'], slug: 'user');
+    addImagePack(client!.accountData['im.ponies.user_emotes'], slug: 'user');
     // next we add all the external image packs
-    final packRooms = client.accountData['im.ponies.emote_rooms'];
+    final packRooms = client!.accountData['im.ponies.emote_rooms'];
     if (packRooms != null && packRooms.content['rooms'] is Map) {
       for (final roomEntry in packRooms.content['rooms'].entries) {
         final roomId = roomEntry.key;
-        final room = client.getRoomById(roomId);
+        final room = client!.getRoomById(roomId);
         if (room != null && roomEntry.value is Map) {
           for (final stateKeyEntry in roomEntry.value.entries) {
             final stateKey = stateKeyEntry.key;
@@ -80,7 +80,7 @@ extension ImagePackRoomExtension on Room {
       for (final entry in allRoomEmotes.entries) {
         addImagePack(entry.value,
             room: this,
-            slug: entry.value.stateKey.isEmpty ? 'room' : entry.value.stateKey);
+            slug: entry.value.stateKey!.isEmpty ? 'room' : entry.value.stateKey);
       }
     }
     return packs;
@@ -88,7 +88,7 @@ extension ImagePackRoomExtension on Room {
 
   /// Get a flat view of all the image packs of a specified [usage], that is a map of all
   /// slugs to a map of the image code to their mxc url
-  Map<String, Map<String, String>> getImagePacksFlat([ImagePackUsage usage]) =>
+  Map<String, Map<String, String>> getImagePacksFlat([ImagePackUsage? usage]) =>
       getImagePacks(usage).map((k, v) =>
           MapEntry(k, v.images.map((k, v) => MapEntry(k, v.url.toString()))));
 }
