@@ -20,125 +20,125 @@ import '../room.dart';
 
 abstract class MatrixLocalizations {
   const MatrixLocalizations();
-  String get emptyChat;
+  String? get emptyChat;
 
-  String get invitedUsersOnly;
+  String? get invitedUsersOnly;
 
-  String get fromTheInvitation;
+  String? get fromTheInvitation;
 
-  String get fromJoining;
+  String? get fromJoining;
 
-  String get visibleForAllParticipants;
+  String? get visibleForAllParticipants;
 
-  String get visibleForEveryone;
+  String? get visibleForEveryone;
 
-  String get guestsCanJoin;
+  String? get guestsCanJoin;
 
-  String get guestsAreForbidden;
+  String? get guestsAreForbidden;
 
-  String get anyoneCanJoin;
+  String? get anyoneCanJoin;
 
   String get needPantalaimonWarning;
 
-  String get channelCorruptedDecryptError;
+  String? get channelCorruptedDecryptError;
 
-  String get encryptionNotEnabled;
+  String? get encryptionNotEnabled;
 
-  String get unknownEncryptionAlgorithm;
+  String? get unknownEncryptionAlgorithm;
 
   String get noPermission;
 
-  String get you;
+  String? get you;
 
-  String get roomHasBeenUpgraded;
+  String? get roomHasBeenUpgraded;
 
-  String groupWith(String displayname);
+  String? groupWith(String? displayname);
 
-  String removedBy(String calcDisplayname);
+  String? removedBy(String? calcDisplayname);
 
-  String sentASticker(String senderName);
+  String? sentASticker(String? senderName);
 
-  String redactedAnEvent(String senderName);
+  String? redactedAnEvent(String? senderName);
 
-  String changedTheRoomAliases(String senderName);
+  String? changedTheRoomAliases(String? senderName);
 
-  String changedTheRoomInvitationLink(String senderName);
+  String? changedTheRoomInvitationLink(String? senderName);
 
-  String createdTheChat(String senderName);
+  String? createdTheChat(String? senderName);
 
-  String changedTheJoinRules(String senderName);
+  String? changedTheJoinRules(String? senderName);
 
-  String changedTheJoinRulesTo(String senderName, String localizedString);
+  String? changedTheJoinRulesTo(String? senderName, String? localizedString);
 
-  String acceptedTheInvitation(String targetName);
+  String? acceptedTheInvitation(String? targetName);
 
-  String rejectedTheInvitation(String targetName);
+  String? rejectedTheInvitation(String? targetName);
 
-  String hasWithdrawnTheInvitationFor(String senderName, String targetName);
+  String? hasWithdrawnTheInvitationFor(String? senderName, String? targetName);
 
-  String joinedTheChat(String targetName);
+  String? joinedTheChat(String? targetName);
 
-  String kickedAndBanned(String senderName, String targetName);
+  String? kickedAndBanned(String? senderName, String? targetName);
 
-  String kicked(String senderName, String targetName);
+  String? kicked(String? senderName, String? targetName);
 
-  String userLeftTheChat(String targetName);
+  String? userLeftTheChat(String? targetName);
 
-  String bannedUser(String senderName, String targetName);
+  String? bannedUser(String? senderName, String? targetName);
 
-  String unbannedUser(String senderName, String targetName);
+  String? unbannedUser(String? senderName, String? targetName);
 
-  String invitedUser(String senderName, String targetName);
+  String? invitedUser(String? senderName, String? targetName);
 
-  String changedTheProfileAvatar(String targetName);
+  String? changedTheProfileAvatar(String? targetName);
 
-  String changedTheDisplaynameTo(String targetName, String newDisplayname);
+  String? changedTheDisplaynameTo(String? targetName, String newDisplayname);
 
-  String changedTheChatPermissions(String senderName);
+  String? changedTheChatPermissions(String? senderName);
 
-  String changedTheChatNameTo(String senderName, String content);
+  String? changedTheChatNameTo(String? senderName, String? content);
 
-  String changedTheChatDescriptionTo(String senderName, String content);
+  String? changedTheChatDescriptionTo(String? senderName, String? content);
 
-  String changedTheChatAvatar(String senderName);
+  String? changedTheChatAvatar(String? senderName);
 
-  String changedTheGuestAccessRules(String senderName);
+  String? changedTheGuestAccessRules(String? senderName);
 
-  String changedTheGuestAccessRulesTo(
-      String senderName, String localizedString);
+  String? changedTheGuestAccessRulesTo(
+      String? senderName, String? localizedString);
 
-  String changedTheHistoryVisibility(String senderName);
+  String? changedTheHistoryVisibility(String? senderName);
 
-  String changedTheHistoryVisibilityTo(
-      String senderName, String localizedString);
+  String? changedTheHistoryVisibilityTo(
+      String? senderName, String? localizedString);
 
-  String activatedEndToEndEncryption(String senderName);
+  String activatedEndToEndEncryption(String? senderName);
 
-  String sentAPicture(String senderName);
+  String? sentAPicture(String? senderName);
 
-  String sentAFile(String senderName);
+  String? sentAFile(String? senderName);
 
-  String sentAnAudio(String senderName);
+  String? sentAnAudio(String? senderName);
 
-  String sentAVideo(String senderName);
+  String? sentAVideo(String? senderName);
 
-  String sharedTheLocation(String senderName);
+  String? sharedTheLocation(String? senderName);
 
-  String couldNotDecryptMessage(String errorText);
+  String? couldNotDecryptMessage(String? errorText);
 
-  String unknownEvent(String typeKey);
+  String? unknownEvent(String typeKey);
 
-  String startedACall(String senderName);
+  String? startedACall(String? senderName);
 
-  String endedTheCall(String senderName);
+  String? endedTheCall(String? senderName);
 
-  String answeredTheCall(String senderName);
+  String? answeredTheCall(String? senderName);
 
-  String sentCallInformations(String senderName);
+  String? sentCallInformations(String? senderName);
 }
 
 extension HistoryVisibilityDisplayString on HistoryVisibility {
-  String getLocalizedString(MatrixLocalizations i18n) {
+  String? getLocalizedString(MatrixLocalizations i18n) {
     switch (this) {
       case HistoryVisibility.invited:
         return i18n.fromTheInvitation;
@@ -154,7 +154,7 @@ extension HistoryVisibilityDisplayString on HistoryVisibility {
 }
 
 extension GuestAccessDisplayString on GuestAccess {
-  String getLocalizedString(MatrixLocalizations i18n) {
+  String? getLocalizedString(MatrixLocalizations i18n) {
     switch (this) {
       case GuestAccess.canJoin:
         return i18n.guestsCanJoin;
@@ -166,7 +166,7 @@ extension GuestAccessDisplayString on GuestAccess {
 }
 
 extension JoinRulesDisplayString on JoinRules {
-  String getLocalizedString(MatrixLocalizations i18n) {
+  String? getLocalizedString(MatrixLocalizations i18n) {
     switch (this) {
       case JoinRules.public:
         return i18n.anyoneCanJoin;
