@@ -32,12 +32,12 @@ void main() {
         ),
       );
       expect(matrixException.errcode, 'M_FORBIDDEN');
-      final flows = matrixException.authenticationFlows;
+      final flows = matrixException.authenticationFlows!;
       expect(flows.length, 1);
       expect(flows.first.stages.length, 1);
       expect(flows.first.stages.first, 'example.type.foo');
       expect(
-        matrixException.authenticationParams['example.type.baz'],
+        matrixException.authenticationParams!['example.type.baz'],
         {'example_key': 'foobar'},
       );
       expect(matrixException.completedAuthenticationFlows.length, 1);
